### 2021-04-24 #3218 @whedon
[PRE REVIEW]: Omnizart: A General Toolbox for Automatic Music Transcription
**Submitting author:** @BreezeWhite (<a href="http://orcid.org/0000-0003-2362-6340">Yu Te Wu</a>)
**Repository:** <a href="https://github.com/Music-and-Culture-Technology-Lab/omnizart" target ="_blank">https://github.com/Music-and-Culture-Technology-Lab/omnizart</a>
**Version:** v0.4.1
