### 2021-04-28 #3218 @Kevin-Mattheus-Moerman
@BreezeWhite thanks for this submission. Our editors in this domain are currently not able to take on this submission. I have therefore added the "waitlisted" label, which means one of our editors will pick up this work once they become available. Thanks for your patience.

In the mean time please consider these comments:

- [ ] Please rename/rephrase the purpose section to be the required `Statement of need` section.
- [ ] Consider more clearly describing the direct scientific research application in the statement of need section
- [ ] We recommend that you shorten the paper where possible, as  [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) state "between 250-1000 words".
