### 2021-04-30 #3218 @Kevin-Mattheus-Moerman
@BreezeWhite if you are at the 1000 word limit that could be fine. As long as you are not repeating content from the documentation. Thanks.
FYI you can update the paper whenever you like here by calling: `@whedon generate pdf`
