### 2021-05-10 fe459f4
Fix passing wrong type of parameters
