+++ omnizart/utils.py
@@ -11,6 +11,7 @@ from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
 
 import jsonschema
 import pretty_midi
+import numpy as np
 import scipy.io.wavfile as wave
 
 from omnizart.constants.midi import SOUNDFONT_PATH
@@ -363,7 +364,7 @@ def aggregate_f0_info(pred, t_unit):
     start_idx = 0
     last_hz = pred[0]
     eps = 1e-6
-    pred.append(0)  # Append an additional zero to the end temporarily.
+    pred = np.append(pred, 0)  # Append an additional zero to the end temporarily.
     while cur_idx < len(pred):
         cur_hz = pred[cur_idx]
         if abs(cur_hz - last_hz) < eps:
@@ -390,5 +391,5 @@ def aggregate_f0_info(pred, t_unit):
         cur_idx += 1
         last_hz = cur_hz
 
-    del pred[-1]  # Remove the additional ending zero.
+    pred = pred[:-1]  # Remove the additional ending zero.
     return results
