+++ omnizart/__init__.py
@@ -7,4 +7,4 @@ SETTING_DIR = os.path.join(MODULE_PATH, "defaults")
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 os.environ['VAMP_PATH'] = os.path.join(MODULE_PATH, "resource", "vamp")
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
