### 2021-05-26 b6252a5
Modify the way of feeding the feature slices with overlapping
