+++ omnizart/music/prediction.py
@@ -29,7 +29,7 @@ def cut_batch_pred(b_pred):
     return np.array(cut_pp)
 
 
-def create_batches(feature, b_size, timesteps, feature_num=384):
+def create_batches_old(feature, b_size, timesteps, feature_num=384):
     frms = np.ceil(len(feature) / timesteps)
     bss = np.ceil(frms / b_size).astype("int")
 
@@ -55,7 +55,127 @@ def create_batches(feature, b_size, timesteps, feature_num=384):
     return batch
 
 
-def predict(feature, model, timesteps=128, feature_num=384, batch_size=4):
+def create_batches(feature, timesteps, b_size=8, step_size=10):
+    """Create a series of batch input.
+
+    The size of the last batch could smaller than the given ``b_size``.
+
+    Parameters
+    ----------
+    feature: numpy.ndarray
+        The only constraint is the first dimension should time index. There is no limit
+        on the number of dimensions.
+    timesteps: int
+        Input feature length of the model.
+    b_size: int
+        Batch size of the input.
+    step_size: int
+        Step size for hopping the feature. Value smaller than ``timesteps`` indicates there
+        will be overlapping between each feature slice.
+
+    Returns
+    -------
+    batches: list
+        List of input batches.
+    """
+    step_size = max(1, min(timesteps, step_size))
+    batches = []
+    batch = []
+    cur_len = 0
+    for idx in range(0, len(feature)-timesteps, step_size):
+        feat = feature[idx:idx+timesteps]
+        batch.append(feat)
+        cur_len += timesteps
+        if len(batch) == b_size:
+            batches.append(batch)
+            batch = []
+
+    feat = feature[cur_len:]
+    if len(feat) < timesteps:
+        pad = np.zeros((timesteps-len(feat),)+feat.shape[1:])
+        feat = np.concatenate([feat, pad])
+
+    if len(batches[-1]) < b_size:
+        batches[-1].append(feat)
+    else:
+        batches.append([feat])
+
+    return batches
+
+
+def merge_batches(batches, step_size=10):
+    """Reverse process of ``create_batches``.
+
+    Merge the list of batch predictions into the complete predicted results.
+
+    Parameters
+    ----------
+    batches: numpy.ndarray
+        List of predicted batches.
+    step_size: int
+        Should be the same as passing to ``create_batches``.
+
+    Returns
+    -------
+    pred: numpy.ndarray
+        The final predicted results.
+    """
+    total_slice = 0
+    for batch in batches:
+        total_slice += len(batch)
+
+    pred_shape = batches[0][0].shape
+    out_len = (total_slice - 1) * step_size + pred_shape[0]
+    output = np.zeros((out_len,) + pred_shape[1:])
+    idx = 0
+    for batch in batches:
+        for pred in batch:
+            start = idx * step_size
+            end = start + pred_shape[0]
+            output[start:end] += pred
+            idx += 1
+
+    mask = np.zeros_like(output)
+    for idx in range(0, out_len-pred_shape[0]+1, step_size):
+        mask[idx:idx+pred_shape[0]] += 1
+    output /= mask
+
+    return output
+
+
+def predict(feature, model, timesteps=128, batch_size=4, step_size=64):
+    """Make predictions on the feature.
+
+    Generate predictions by using the loaded model.
+
+    Parameters
+    ----------
+    feature: numpy.ndarray
+        Extracted feature of the audio. Dimension: timesteps x feature_size x channels
+    model: keras.Model
+        The loaded model instance.
+    batch_size: int
+        Batch size for the prediction iteration.
+    step_size: int
+        Step size for hopping the feature. Value smaller then ``timesteps`` means there will be
+        overlapping.
+
[…]
