### 2021-05-26 fac2d2e
Minor updates of peak finding parameters
