### 2021-05-26 afbc09f
Add value of prediction's entropy to smooth loss
