### 2021-05-26 90d60ea
Modify training loop to use the built-in model.fit.
