+++ omnizart/music/app.py
@@ -111,12 +111,7 @@ class MusicTranscription(BaseTranscription):
 
         logger.info("Predicting...")
         channels = [FEATURE_NAME_TO_NUMBER[ch_name] for ch_name in model_settings.training.channels]
-        pred = predict(
-            feature[:, :, channels],
-            model,
-            timesteps=model_settings.training.timesteps,
-            feature_num=model_settings.training.feature_num
-        )
+        pred = predict(feature[:, :, channels], model, timesteps=model_settings.training.timesteps)
 
         logger.info("Infering notes....")
         midi = multi_inst_note_inference(
@@ -132,7 +127,7 @@ class MusicTranscription(BaseTranscription):
 
         self._output_midi(output=output, input_audio=input_audio, midi=midi)
         if os.environ.get("LOG_LEVEL", "") == "debug":
-            dump_pickle({"pred": pred}, "./debug_pred.pickle")
+            dump_pickle({"pred": pred, "feature": feature}, "./debug_pred.pickle")
 
         logger.info("Transcription finished")
         return midi
@@ -272,24 +267,28 @@ class MusicTranscription(BaseTranscription):
         train_feat_files, val_feat_files = get_train_val_feat_file_list(feature_folder, split=split)
 
         output_types = (tf.float32, tf.float32)
+        output_shapes = (
+            (settings.training.timesteps, settings.training.feature_num, len(settings.training.channels)),
+            (settings.training.timesteps, settings.training.feature_num, l_type.get_out_classes())
+        )
         train_dataset = MusicDatasetLoader(
                 l_type.get_conversion_func(),
                 feature_files=train_feat_files,
-                num_samples=settings.training.batch_size * settings.training.steps,
+                num_samples=settings.training.epoch * settings.training.batch_size * settings.training.steps,
                 timesteps=settings.training.timesteps,
                 channels=[FEATURE_NAME_TO_NUMBER[ch_name] for ch_name in settings.training.channels],
                 feature_num=settings.training.feature_num
             ) \
-            .get_dataset(settings.training.batch_size, output_types=output_types)
+            .get_dataset(settings.training.batch_size, output_types=output_types, output_shapes=output_shapes)
         val_dataset = MusicDatasetLoader(
                 l_type.get_conversion_func(),
                 feature_files=val_feat_files,
-                num_samples=settings.training.val_batch_size * settings.training.val_steps,
+                num_samples=settings.training.epoch * settings.training.val_batch_size * settings.training.val_steps,
                 timesteps=settings.training.timesteps,
                 channels=[FEATURE_NAME_TO_NUMBER[ch_name] for ch_name in settings.training.channels],
                 feature_num=settings.training.feature_num
             ) \
-            .get_dataset(settings.training.val_batch_size, output_types=output_types)
+            .get_dataset(settings.training.val_batch_size, output_types=output_types, output_shapes=output_shapes)
 
         if input_model_path is None:
             logger.info("Creating new model with type: %s", settings.model.model_type)
@@ -305,11 +304,12 @@ class MusicTranscription(BaseTranscription):
 
         logger.info("Compiling model with loss function type: %s", settings.training.loss_function)
         loss_func = {
-            "smooth": lambda y, x: smooth_loss(y, x, total_chs=l_type.get_out_classes()),
+            "smooth": lambda y, x: smooth_loss(y, x, total_chs=l_type.get_out_classes(), weight=[1, 0.5, 1.5]),
             "focal": focal_loss,
             "bce": tf.keras.losses.BinaryCrossentropy()
         }[settings.training.loss_function]
[…]
