### 2021-05-28 d795b4f
Remove entropy value from smooth loss
