### 2021-05-28 6140a2f
Add learning rate scheduler callback
