### 2021-05-28 3481ef5
Fix a critical bug that could lead to missing onsets.
