### 2021-05-31 58897e1
Modify default onset threshold of music piano model
