### 2021-05-31 748fb17
Various improvements on infering notes
