+++ setup.py
@@ -59,11 +59,11 @@ entry_points = \
 
 setup_kwargs = {
     'name': 'omnizart',
-    'version': '0.3.4',
+    'version': '0.4.0',
     'description': 'Omniscient Mozart, being able to transcribe everything in the music.',
     'long_description': '# OMNIZART\n\n[![build](https://github.com/Music-and-Culture-Technology-Lab/omnizart/workflows/general-check/badge.svg)](https://github.com/Music-and-Culture-Technology-Lab/omnizart/actions?query=workflow%3Ageneral-check)\n[![docs](https://github.com/Music-and-Culture-Technology-Lab/omnizart/workflows/docs/badge.svg?branch=build_doc)](https://music-and-culture-technology-lab.github.io/omnizart-doc/)\n[![PyPI version](https://badge.fury.io/py/omnizart.svg)](https://badge.fury.io/py/omnizart)\n![PyPI - License](https://img.shields.io/pypi/l/omnizart)\n[![PyPI - Downloads](https://img.shields.io/pypi/dm/omnizart)](https://pypistats.org/packages/omnizart)\n[![Docker Pulls](https://img.shields.io/docker/pulls/mctlab/omnizart)](https://hub.docker.com/r/mctlab/omnizart)\n\nOmnizart is a Python library that aims for democratizing automatic music transcription.\nGiven polyphonic music, it is able to transcribe pitched instruments, vocal melody, chords, drum events, and beat.\nThis is powered by the research outcomes from [Music and Culture Technology (MCT) Lab](https://sites.google.com/view/mctl/home).\n\n### Transcribe your favorite songs now in Colab! [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://bit.ly/OmnizartColab)\n\n# Quick start\n\nVisit the [complete document](https://music-and-culture-technology-lab.github.io/omnizart-doc/) for detailed guidance.\n\n## Pip\n``` bash\n# Install omnizart\npip install omnizart\n\n# Download the checkpoints\nomnizart download-checkpoints\n\n# Transcribe your songs\nomnizart drum transcribe <path/to/audio.wav>\nomnizart chord transcribe <path/to/audio.wav>\nomnizart music transcribe <path/to/audio.wav>\n```\n\n## Docker\n```\ndocker pull mctlab/omnizart:latest\ndocker run -it mctlab/omnizart:latest bash\n```\n\n# Supported applications\n| Application      | Transcription      | Training           | Evaluation | Description                                      |\n|------------------|--------------------|--------------------|------------|--------------------------------------------------|\n| music            | :heavy_check_mark: | :heavy_check_mark: |            | Transcribe musical notes of pitched instruments. |\n| drum             | :heavy_check_mark: | :interrobang:      |            | Transcribe events of percussive instruments.     |\n| vocal            | :heavy_check_mark: | :heavy_check_mark: |            | Transcribe note-level vocal melody.              |\n| vocal-contour    | :heavy_check_mark: | :heavy_check_mark: |            | Transcribe frame-level vocal melody (F0).        |\n| chord            | :heavy_check_mark: | :heavy_check_mark: |            | Transcribe chord progressions.                   |\n| beat             | :heavy_check_mark: | :heavy_check_mark: |            | Transcribe beat position.                        |\n\n**NOTES**\nThe current implementation for the drum model has unknown bugs, preventing loss convergence when training from scratch.\nFortunately, you can still enjoy drum transcription with the provided checkpoints.\n\n',
     'author': 'BreezeWhite',
-    'author_email': 'freedombluewater@gmail.com',
+    'author_email': 'miyashita2010@tuta.io',
     'maintainer': None,
     'maintainer_email': None,
     'url': 'https://sites.google.com/view/mctl/home',
