+++ pyproject.toml
@@ -1,9 +1,9 @@
 [tool.poetry]
 name = "omnizart"
-version = "0.3.4"
+version = "0.4.0"
 description = "Omniscient Mozart, being able to transcribe everything in the music."
 readme = "README.md"
-authors = ["BreezeWhite <freedombluewater@gmail.com>", "yjlolo <yjlolo125@gmail.com>"]
+authors = ["BreezeWhite <miyashita2010@tuta.io>", "yjlolo <yjlolo125@gmail.com>"]
 license = "MIT"
 repository = "https://github.com/Music-and-Culture-Technology-Lab/omnizart"
 documentation = "https://music-and-culture-technology-lab.github.io/omnizart-doc/"
