### 2021-06-01 #3218 @arfon
:wave: @faroit - this looks like it could be perfect for you as a first paper?
