### 2021-06-03 5ec262b
Select the newest weight file if there are multiple
