### 2021-06-03 848b070
Fix bug of parsing weight files
