+++ omnizart/checkpoints/music/music_piano-v2/arch.yaml
@@ -1,0 +1,5567 @@
+backend: tensorflow
+class_name: Functional
+config:
+  input_layers:
+  - - input_score_48
+    - 0
+    - 0
+  layers:
+  - class_name: InputLayer
+    config:
+      batch_input_shape: !!python/tuple
+      - null
+      - 256
+      - 352
+      - 3
+      dtype: float32
+      name: input_score_48
+      ragged: false
+      sparse: false
+    inbound_nodes: []
+    name: input_score_48
+  - class_name: Conv2D
+    config:
+      activation: linear
+      activity_regularizer: null
+      bias_constraint: null
+      bias_initializer:
+        class_name: Zeros
+        config: {}
+      bias_regularizer: null
+      data_format: channels_last
+      dilation_rate: &id002 !!python/tuple
+      - 1
+      - 1
+      dtype: float32
+      filters: 128
+      groups: 1
+      kernel_constraint: null
+      kernel_initializer:
+        class_name: GlorotUniform
+        config:
+          seed: null
+      kernel_regularizer: null
+      kernel_size: !!python/tuple
+      - 7
+      - 7
+      name: conv2d
+      padding: same
+      strides: !!python/tuple
+      - 1
+      - 1
+      trainable: true
+      use_bias: true
+    inbound_nodes:
+    - - - input_score_48
+        - 0
+        - 0
+        - {}
+    name: conv2d
+  - class_name: Activation
+    config:
+      activation: relu
+      dtype: float32
+      name: activation
+      trainable: true
+    inbound_nodes:
+    - - - conv2d
+        - 0
+        - 0
+        - {}
+    name: activation
+  - class_name: BatchNormalization
+    config:
+      axis: !!python/object/apply:tensorflow.python.training.tracking.data_structures.ListWrapper
+      - - 3
+      beta_constraint: null
+      beta_initializer:
+        class_name: Zeros
+        config: {}
+      beta_regularizer: null
+      center: true
+      dtype: float32
+      epsilon: 0.001
+      gamma_constraint: null
+      gamma_initializer:
+        class_name: Ones
+        config: {}
+      gamma_regularizer: null
+      momentum: 0.99
+      moving_mean_initializer:
+        class_name: Zeros
+        config: {}
+      moving_variance_initializer:
+        class_name: Ones
+        config: {}
+      name: batch_normalization
+      scale: true
+      trainable: true
+    inbound_nodes:
+    - - - activation
+        - 0
+        - 0
+        - {}
+    name: batch_normalization
+  - class_name: Dropout
+    config:
+      dtype: float32
+      name: dropout
+      noise_shape: null
+      rate: 0.4
+      seed: null
+      trainable: true
+    inbound_nodes:
+    - - - batch_normalization
+        - 0
+        - 0
+        - {}
+    name: dropout
+  - class_name: Conv2D
+    config:
+      activation: linear
+      activity_regularizer: null
+      bias_constraint: null
+      bias_initializer:
+        class_name: Zeros
+        config: {}
+      bias_regularizer: null
+      data_format: channels_last
+      dilation_rate: !!python/tuple
+      - 1
+      - 1
+      dtype: float32
+      filters: 32
+      groups: 1
+      kernel_constraint: null
+      kernel_initializer:
+        class_name: GlorotUniform
+        config:
+          seed: null
+      kernel_regularizer: null
+      kernel_size: &id001 !!python/tuple
+      - 3
+      - 3
+      name: conv2d_1
+      padding: same
+      strides: &id003 !!python/tuple
+      - 2
+      - 2
+      trainable: true
+      use_bias: true
+    inbound_nodes:
+    - - - dropout
+        - 0
+        - 0
+        - {}
+    name: conv2d_1
[…]
