### 2021-06-04 #3218 @faroit
👋  Hi @BreezeWhite, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

As this toolbox implements a bunch of transcription models, it would be great to have reviewers with domain knowledge who also have experience with frameworks used in the MIR community... I have a few potential reviewers in mind:

- __@ superbock__: author of [madmom](https://github.com/CPJKU/madmom)
- __@ adarob:__ [magenta](https://github.com/magenta/magenta) core dev
- __@ keunwoochoi:__ author of [DrummerNet](https://github.com/keunwoochoi/DrummerNet)

@BreezeWhite, I am also interested in your input on potential reviewers, so any suggestions you have are also welcome.
For this, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
