+++ paper.bib
@@ -13,7 +13,8 @@ journal={IEEE/ACM Transactions on Audio, Speech, and Language Processing},
 volume={28},
 pages={2796--2809},
 year={2020},
-publisher={IEEE}
+publisher={IEEE},
+doi={10.1109/taslp.2020.3030482}
 }
 
 @inproceedings{hawthorne2018enabling,
@@ -51,6 +52,7 @@ title={Encoder-Decoder with Atrous Separable Convolution for Semantic Image Segm
 author={Chen, L.-C. and Zhu, Y. and P. George and S. Florian and A. Hartwig},
 journal={arXiv preprint arXiv:1802.02611},
 year={2018},
+doi={10.1007/978-3-030-01234-2_49}
 }
 
 @article{su2015combining,  
@@ -61,7 +63,8 @@ year={2015},
 volume={23},  
 number={10},  
 pages={1600-1612},
-publisher={IEEE}
+publisher={IEEE},
+doi={10.1109/taslp.2015.2442411}
 }
 
 @inproceedings{kelz2016potential,
@@ -82,6 +85,7 @@ title={Improving automatic drum transcription using large-scale audio-to-midi al
 author={Wei, I-Chieh and Wu, Chih-Wei and Su, Li},
 booktitle={Proc. ICASSP},
 year={2021},
+doi={10.1109/icassp39728.2021.9414409}
 }
 
 @inproceedings{wu2018automatic,
@@ -89,7 +93,8 @@ title={Automatic {M}usic {T}ranscription {L}everaging {G}eneralized {C}epstral {
 author={Wu, Yu-Te and Chen, Berlin and Su, Li},
 booktitle={{IEEE} International Conference on Acoustics, Speech and Signal Processing ({ICASSP})},
 pages={401--405},
-year={2018}
+year={2018},
+doi={10.1109/icassp.2018.8462079}
 }
 
 @INPROCEEDINGS{su2018vocal, 
@@ -97,7 +102,8 @@ author={Su, L.},
 booktitle={Proc. ICASSP}, 
 title={Vocal Melody Extraction Using Patch-Based {CNN}}, 
 year={2018}, 
-pages={371-375}
+pages={371-375},
+doi={10.1109/icassp.2018.8462420}
 }
 
 @inproceedings{fu2019hierarchical,
@@ -114,7 +120,8 @@ journal={IEEE Access},
 volume={7},
 pages={186126--186136},
 year={2019},
-publisher={IEEE}
+publisher={IEEE},
+doi={10.1109/access.2019.2960566}
 }
 
 @article{miyato2018virtual,
@@ -125,7 +132,8 @@ volume={41},
 number={8},
 pages={1979--1993},
 year={2018},
-publisher={IEEE}
+publisher={IEEE},
+doi={10.1109/tpami.2018.2858821}
 }
 
 @inproceedings{chen2019harmony,
@@ -151,7 +159,8 @@ volume={18},
 number={2},
 pages={310--319},
 year={2009},
-publisher={IEEE}
+publisher={IEEE},
+doi={10.1109/tasl.2009.2026503}
 }
 
 @inproceedings{molina2014evaluation,
@@ -192,7 +201,8 @@ author={John Thickstun and
         Sham M. Kakade},
 booktitle={{IEEE} International Conference on Acoustics, Speech and Signal Processing ({ICASSP})},
 pages={2241--2245},
-year={2018}
+year={2018},
+doi={10.1109/icassp.2018.8461686}
 }
 
 @inproceedings{bock2016madmom,
@@ -210,7 +220,8 @@ booktitle={Proceedings of the 14th python in science conference},
 volume={8},
 pages={18--25},
 year={2015},
[…]
