### 2021-06-05 #3218 @BreezeWhite
Hi @faroit, thank you for the suggestions.

For the DOI issue, I've just updated the missing DOIs in the `paper.bib`, you could check if there are still missing things.
And for the reviewers, I would like to mention @ bmfcc for joining the review process. And also thank you for mentioning the above members that would be a huge help for reviewing the paper.
