### 2021-06-06 #3218 @faroit
@BreezeWhite

> And for the reviewers, I would like to mention **@ bmfcc** for joining the review process. And also thank you for mentioning the above members that would be a huge help for reviewing the paper.

by @ bmfcc ... were you referring to @ bmcfee ? In that case, I'm afraid that he is to busy and is not available for a review as he also serves as an topic editor at JOSS.
