### 2021-06-14 #3218 @faroit
@Kevin-Mattheus-Moerman Is there anything left in the comments you made in https://github.com/openjournals/joss-reviews/issues/3218#issuecomment-828230199 that should be addressed by @BreezeWhite before we proceed with the review process?
