### 2021-06-14 #3218 @Kevin-Mattheus-Moerman
@faroit thanks for letting me check this before moving on. Yes I ticked those boxes.

I have one more minor comment/recommendation below.

I recommend that the following sentence is rephrased:
`We believe the release of Omnizart can accelerate the advance of AMT research and contribute to the MIR community.`.

I have 2 issues with this sentence.

1. I would avoid statements like "We" or "I" in these papers. Furthermore, the statement `We believe..`, makes it feel like "only a belief", i.e. that there isn't necessarily clear/good evidence to support it. Could the authors rephrase the sentence, if they have confidence in their claims, such that it sounds more firmly grounded in evidence?
2. The statement `..contribute to the MIR community.` sounds vague, it is unclear what it contributes. Perhaps rephrase.

I'm not saying the following is the best possible sentence (give this some thought and formulate something yourself) but the following would be an example to avoid the above issues (perhaps summarize something in place of those dots):
`Omnizart is an important tool for the MIR community, since it can accelerate and advance AMT research by.....`

@faroit I'll leave you to follow up with this if that is okay, and if you agree.
