### 2021-06-17 c3ed96c
Rephrase paper sentences for JOSS submission
