### 2021-06-19 Music-and-Culture-Technology-Lab/omnizart#25 @keunwoochoi
My comments on `paper.md`
Hi, I'm one of the JOSS reviewers of this repo. Before checking the software side of the work, I'd like to leave my comments on the paper here. I read this version (https://github.com/openjournals/joss-reviews/issues/3218#issuecomment-855172230) of the paper. I noticed it's a bit updated but not the part I'd like to mention about.

---

`L11-14`: It's a bit long, making the sentence slightly vague.

`L25-26`: What do you mean exactly by "multifaceted nature"? This can be clarified.

`L29`: "previous solution" -> "previous solutions". Also, maybe some citations of the solutions?

`L30-31`: I think this sentences can be re-written a bit to clarify. The current one confused me a bit.

`L34-35`: I found this sentence out of context a little. Wouldn't it make more sense to mention it in the first paragraph of "Statement of need"?

`L36-37`: -> "..original papers. The models can.."

`L56-58`: Really? But how much instrument-agnostic is it?

`L61-62`: "which predicts percussive events from a given input audio." → probably no need to mention? This is assumed in the definition of drum transcription.

`L69-70`: It sounds like the pitch extractor not a "module" while note segmentation is, and that confuses me.

`L76` : "The training includes.." → "The training data includes.."

`L80`: "function" → probably, keep using "model" for consistency?

`L109`: remove "also".

---

Best,
Keunwoo
