### 2021-06-21 #3391 @faroit
👏 @hagenw @keunwoochoi - This is the review thread for the paper. All of our communications will happen here from now on.

Both of you have checklists at the top of this thread (in that first comment) with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
