### 2021-06-21 Music-and-Culture-Technology-Lab/omnizart#25 @faroit
just for the referencing: the review issue is here: https://github.com/openjournals/joss-reviews/issues/3347
