### 2021-06-21 Music-and-Culture-Technology-Lab/omnizart#25 @keunwoochoi
> just for the referencing: the review issue is here: [openjournals/joss-reviews#3347](https://github.com/openjournals/joss-reviews/issues/3347)

oops, i think @faroit  you meant "https://github.com/openjournals/joss-reviews/issues/3391" this. but you mean, comments like this are supposed to be communicated at the review thread?
