### 2021-06-21 Music-and-Culture-Technology-Lab/omnizart#25 @faroit
@keunwoochoi yes, sure, sorry. copy-pasted the wrong link....

> but you mean, comments like this are supposed to be communicated at the review thread?

No, its exactly how it's supposed to be :+1:
I was just linking to the review issue, since you started reviewing before the official review even started ;-)
