### 2021-06-27 Music-and-Culture-Technology-Lab/omnizart#26 @keunwoochoi
Installation issues on MacOS
(This is a part of JOSS review https://github.com/openjournals/joss-reviews/issues/3391)

Hi, I had three issues installing omnizart on macos using `pip install omnizart` (python3.9, miniconda).

The first issue is coming from `madmom` which asks to install `Cython` manually. This is technically not the issue of `omnizart`, but it'd be nice if this is mentioned somewhere for beginners.

The second issue is coming from `vamp`.
```
Collecting vamp<2.0.0,>=1.1.0
  Using cached vamp-1.1.0.tar.gz (107 kB)
     command: /Users/keunwoochoi/miniconda3/envs/omnizart/bin/python -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-bk7368ox/vamp_76483087bd584264b4ab165b4152fe3a/setup.py'"'"'; __file__='"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-bk7368ox/vamp_76483087bd584264b4ab165b4152fe3a/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-pip-egg-info-d6aulwnk
         cwd: /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-bk7368ox/vamp_76483087bd584264b4ab165b4152fe3a/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-bk7368ox/vamp_76483087bd584264b4ab165b4152fe3a/setup.py", line 3, in <module>
        import numpy as np
    ModuleNotFoundError: No module named 'numpy'
WARNING: Discarding https://files.pythonhosted.org/packages/66/65/a54958777fa667ef881ceec18dd0b423b2ba2c77543060b21aff0e1dbcb6/vamp-1.1.0.tar.gz#sha256=d40e0916e958e2806d37920f903c2518cc46959a4f21e020737b6a48f6e7e82d (from https://pypi.org/simple/vamp/). Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
ERROR: Could not find a version that satisfies the requirement vamp<2.0.0,>=1.1.0 (from omnizart) (from versions: 1.0.0, 1.0.1, 1.0.2, 1.1.0)
ERROR: No matching distribution found for vamp<2.0.0,>=1.1.0
```

I could avoid this issue by manually installing `numpy`. This is probably, again, due to `vamp` not `omnizart`, but maybe this should be mentioned somewhere as well? Moreover, since `numpy` is used in `omnizart` directly (https://github.com/Music-and-Culture-Technology-Lab/omnizart/search?q=import+numpy), shouldn't the `setup.py` include `numpy` as a requirement? That would also solve the installation issue anyway so maybe not a bad idea.

Finally, the third issue is about `numba` and `llvmlite`.
```
(omnizart) keunwoochoi@KCs-qmul-mbp Downloads $ pip install numpy
Collecting numpy
  Downloading numpy-1.21.0-cp39-cp39-macosx_10_9_x86_64.whl (16.9 MB)
     |████████████████████████████████| 16.9 MB 900 kB/s
Installing collected packages: numpy
Successfully installed numpy-1.21.0
(omnizart) keunwoochoi@KCs-qmul-mbp Downloads $ pip install omnizart
Collecting omnizart
  Using cached omnizart-0.1.0-py3-none-any.whl (3.5 MB)
Collecting tensorflow<3.0.0,>=2.3.0
  Downloading tensorflow-2.5.0-cp39-cp39-macosx_10_11_x86_64.whl (195.7 MB)
     |████████████████████████████████| 195.7 MB 109 kB/s
Collecting tqdm<5.0.0,>=4.49.0
  Using cached tqdm-4.61.1-py2.py3-none-any.whl (75 kB)
Collecting pyfluidsynth<2.0.0,>=1.2.5
  Using cached pyFluidSynth-1.3.0-py3-none-any.whl (18 kB)
Collecting madmom<0.17.0,>=0.16.1
  Using cached madmom-0.16.1.tar.gz (20.0 MB)
Collecting jsonschema<4.0.0,>=3.2.0
  Using cached jsonschema-3.2.0-py2.py3-none-any.whl (56 kB)
Collecting numba==0.48
  Downloading numba-0.48.0.tar.gz (2.0 MB)
     |████████████████████████████████| 2.0 MB 7.6 MB/s
Collecting opencv-python<5.0.0,>=4.4.0
  Using cached opencv_python-4.5.2.54-cp39-cp39-macosx_10_15_x86_64.whl (43.7 MB)
Collecting urllib3<2.0.0,>=1.25.11
  Using cached urllib3-1.26.6-py2.py3-none-any.whl (138 kB)
Collecting pyyaml<6.0.0,>=5.3.1
  Downloading PyYAML-5.4.1-cp39-cp39-macosx_10_9_x86_64.whl (259 kB)
     |████████████████████████████████| 259 kB 4.1 MB/s
Collecting click<8.0.0,>=7.1.2
  Using cached click-7.1.2-py2.py3-none-any.whl (82 kB)
Collecting vamp<2.0.0,>=1.1.0
  Using cached vamp-1.1.0.tar.gz (107 kB)
Collecting pretty_midi<0.3.0,>=0.2.9
  Using cached pretty_midi-0.2.9.tar.gz (5.6 MB)
Collecting librosa<0.9.0,>=0.8.0
  Using cached librosa-0.8.1-py3-none-any.whl (203 kB)
Collecting llvmlite<0.32.0,>=0.31.0dev0
  Downloading llvmlite-0.31.0.tar.gz (110 kB)
     |████████████████████████████████| 110 kB 5.7 MB/s
Requirement already satisfied: numpy>=1.15 in /Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/site-packages (from numba==0.48->omnizart) (1.21.0)
Requirement already satisfied: setuptools in /Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/site-packages (from numba==0.48->omnizart) (52.0.0.post20210125)
Collecting pyrsistent>=0.14.0
  Using cached pyrsistent-0.17.3.tar.gz (106 kB)
Collecting six>=1.11.0
  Downloading six-1.16.0-py2.py3-none-any.whl (11 kB)
Collecting attrs>=17.4.0
  Downloading attrs-21.2.0-py2.py3-none-any.whl (53 kB)
     |████████████████████████████████| 53 kB 2.8 MB/s
Collecting packaging>=20.0
  Downloading packaging-20.9-py2.py3-none-any.whl (40 kB)
     |████████████████████████████████| 40 kB 1.4 MB/s
Collecting resampy>=0.2.2
  Using cached resampy-0.2.2.tar.gz (323 kB)
Collecting decorator>=3.0.0
  Downloading decorator-5.0.9-py3-none-any.whl (8.9 kB)
Collecting pooch>=1.0
  Using cached pooch-1.4.0-py3-none-any.whl (51 kB)
Collecting joblib>=0.14
  Downloading joblib-1.0.1-py3-none-any.whl (303 kB)
     |████████████████████████████████| 303 kB 3.9 MB/s
Collecting scikit-learn!=0.19.0,>=0.14.0
  Downloading scikit_learn-0.24.2-cp39-cp39-macosx_10_13_x86_64.whl (7.3 MB)
     |████████████████████████████████| 7.3 MB 6.6 MB/s
Collecting soundfile>=0.10.2
  Using cached SoundFile-0.10.3.post1-py2.py3.cp26.cp27.cp32.cp33.cp34.cp35.cp36.pp27.pp32.pp33-none-macosx_10_5_x86_64.macosx_10_6_intel.macosx_10_9_intel.macosx_10_9_x86_64.whl (613 kB)
Collecting scipy>=1.0.0
  Downloading scipy-1.7.0-cp39-cp39-macosx_10_9_x86_64.whl (32.1 MB)
     |████████████████████████████████| 32.1 MB 3.1 MB/s
Collecting audioread>=2.0.0
  Downloading audioread-2.1.9.tar.gz (377 kB)
     |████████████████████████████████| 377 kB 4.1 MB/s
Requirement already satisfied: cython>=0.25 in /Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/site-packages (from madmom<0.17.0,>=0.16.1->omnizart) (0.29.23)
Collecting mido>=1.2.8
  Using cached mido-1.2.10-py2.py3-none-any.whl (51 kB)
Collecting pyparsing>=2.0.2
  Using cached pyparsing-2.4.7-py2.py3-none-any.whl (67 kB)
Collecting requests
  Using cached requests-2.25.1-py2.py3-none-any.whl (61 kB)
Collecting appdirs
  Using cached appdirs-1.4.4-py2.py3-none-any.whl (9.6 kB)
Collecting threadpoolctl>=2.0.0
  Using cached threadpoolctl-2.1.0-py3-none-any.whl (12 kB)
Collecting cffi>=1.0
  Downloading cffi-1.14.5-cp39-cp39-macosx_10_9_x86_64.whl (177 kB)
     |████████████████████████████████| 177 kB 3.8 MB/s
Collecting pycparser
  Using cached pycparser-2.20-py2.py3-none-any.whl (112 kB)
Collecting protobuf>=3.9.2
  Downloading protobuf-3.17.3-cp39-cp39-macosx_10_9_x86_64.whl (1.0 MB)
     |████████████████████████████████| 1.0 MB 4.7 MB/s
Collecting keras-preprocessing~=1.1.2
  Using cached Keras_Preprocessing-1.1.2-py2.py3-none-any.whl (42 kB)
Collecting astunparse~=1.6.3
  Using cached astunparse-1.6.3-py2.py3-none-any.whl (12 kB)
Collecting opt-einsum~=3.3.0
  Using cached opt_einsum-3.3.0-py3-none-any.whl (65 kB)
Collecting six>=1.11.0
  Using cached six-1.15.0-py2.py3-none-any.whl (10 kB)
Collecting typing-extensions~=3.7.4
  Using cached typing_extensions-3.7.4.3-py3-none-any.whl (22 kB)
Collecting numpy>=1.15
  Downloading numpy-1.19.5-cp39-cp39-macosx_10_9_x86_64.whl (15.6 MB)
     |████████████████████████████████| 15.6 MB 2.2 MB/s
Collecting tensorflow-estimator<2.6.0,>=2.5.0rc0
  Using cached tensorflow_estimator-2.5.0-py2.py3-none-any.whl (462 kB)
Collecting tensorboard~=2.5
  Using cached tensorboard-2.5.0-py3-none-any.whl (6.0 MB)
Collecting google-pasta~=0.2
  Using cached google_pasta-0.2.0-py3-none-any.whl (57 kB)
Collecting grpcio~=1.34.0
  Downloading grpcio-1.34.1-cp39-cp39-macosx_10_10_x86_64.whl (3.7 MB)
     |████████████████████████████████| 3.7 MB 3.4 MB/s
Collecting keras-nightly~=2.5.0.dev
  Using cached keras_nightly-2.5.0.dev2021032900-py2.py3-none-any.whl (1.2 MB)
Collecting h5py~=3.1.0
  Downloading h5py-3.1.0-cp39-cp39-macosx_10_9_x86_64.whl (2.9 MB)
     |████████████████████████████████| 2.9 MB 6.0 MB/s
Requirement already satisfied: wheel~=0.35 in /Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/site-packages (from tensorflow<3.0.0,>=2.3.0->omnizart) (0.36.2)
Collecting wrapt~=1.12.1
  Using cached wrapt-1.12.1.tar.gz (27 kB)
Collecting termcolor~=1.1.0
  Using cached termcolor-1.1.0.tar.gz (3.9 kB)
Collecting flatbuffers~=1.12.0
  Using cached flatbuffers-1.12-py2.py3-none-any.whl (15 kB)
Collecting absl-py~=0.10
  Downloading absl_py-0.13.0-py3-none-any.whl (132 kB)
     |████████████████████████████████| 132 kB 4.4 MB/s
Collecting gast==0.4.0
  Using cached gast-0.4.0-py3-none-any.whl (9.8 kB)
Collecting google-auth-oauthlib<0.5,>=0.4.1
  Using cached google_auth_oauthlib-0.4.4-py2.py3-none-any.whl (18 kB)
Collecting google-auth<2,>=1.6.3
  Downloading google_auth-1.32.0-py2.py3-none-any.whl (147 kB)
     |████████████████████████████████| 147 kB 7.4 MB/s
Collecting werkzeug>=0.11.15
  Using cached Werkzeug-2.0.1-py3-none-any.whl (288 kB)
Collecting tensorboard-plugin-wit>=1.6.0
  Using cached tensorboard_plugin_wit-1.8.0-py3-none-any.whl (781 kB)
Collecting markdown>=2.6.8
  Using cached Markdown-3.3.4-py3-none-any.whl (97 kB)
Collecting tensorboard-data-server<0.7.0,>=0.6.0
  Using cached tensorboard_data_server-0.6.1-py3-none-macosx_10_9_x86_64.whl (3.5 MB)
Collecting cachetools<5.0,>=2.0.0
  Using cached cachetools-4.2.2-py3-none-any.whl (11 kB)
Collecting pyasn1-modules>=0.2.1
  Using cached pyasn1_modules-0.2.8-py2.py3-none-any.whl (155 kB)
Collecting rsa<5,>=3.1.4
  Using cached rsa-4.7.2-py3-none-any.whl (34 kB)
Collecting requests-oauthlib>=0.7.0
  Using cached requests_oauthlib-1.3.0-py2.py3-none-any.whl (23 kB)
Collecting pyasn1<0.5.0,>=0.4.6
  Using cached pyasn1-0.4.8-py2.py3-none-any.whl (77 kB)
Collecting idna<3,>=2.5
  Using cached idna-2.10-py2.py3-none-any.whl (58 kB)
Collecting chardet<5,>=3.0.2
  Using cached chardet-4.0.0-py2.py3-none-any.whl (178 kB)
Requirement already satisfied: certifi>=2017.4.17 in /Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/site-packages (from requests->pooch>=1.0->librosa<0.9.0,>=0.8.0->omnizart) (2021.5.30)
Collecting oauthlib>=3.0.0
  Using cached oauthlib-3.1.1-py2.py3-none-any.whl (146 kB)
Building wheels for collected packages: numba, audioread, llvmlite, madmom, pretty-midi, pyrsistent, resampy, termcolor, vamp, wrapt
  Building wheel for numba (setup.py) ... done
  Created wheel for numba: filename=numba-0.48.0-cp39-cp39-macosx_10_15_x86_64.whl size=2023786 sha256=2f2d94926e7a7fad8b499957b533d5643c8b6fc25d48a8663839ca8d5bdb4073
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/67/f4/34/43855bda1e661dc1afedcb295d313a510239d901336fd636c3
  Building wheel for audioread (setup.py) ... done
  Created wheel for audioread: filename=audioread-2.1.9-py3-none-any.whl size=23141 sha256=e05ddce16154efe95cce8eff92ada4e631b6fd68b20562490b011eb73e0cc5a8
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/a2/a3/bd/ec1568ce7515115a11ab686d509ad302124c782af065de47ee
  Building wheel for llvmlite (setup.py) ... error
   command: /Users/keunwoochoi/miniconda3/envs/omnizart/bin/python -u -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/setup.py'"'"'; __file__='"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-wheel-iow1oe63
       cwd: /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/
  Complete output (26 lines):
  running bdist_wheel
  /Users/keunwoochoi/miniconda3/envs/omnizart/bin/python /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/ffi/build.py
  LLVM version... Traceback (most recent call last):
    File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/ffi/build.py", line 105, in main_posix
      out = subprocess.check_output([llvm_config, '--version'])
    File "/Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/subprocess.py", line 424, in check_output
      return run(*popenargs, stdout=PIPE, timeout=timeout, check=True,
    File "/Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/subprocess.py", line 505, in run
      with Popen(*popenargs, **kwargs) as process:
    File "/Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/subprocess.py", line 951, in __init__
      self._execute_child(args, executable, preexec_fn, close_fds,
    File "/Users/keunwoochoi/miniconda3/envs/omnizart/lib/python3.9/subprocess.py", line 1821, in _execute_child
      raise child_exception_type(errno_num, err_msg, err_filename)
  FileNotFoundError: [Errno 2] No such file or directory: 'llvm-config'

  Traceback (most recent call last):
    File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/ffi/build.py", line 168, in <module>
      main()
    File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/ffi/build.py", line 162, in main
      main_posix('osx', '.dylib')
    File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/ffi/build.py", line 107, in main_posix
      raise RuntimeError("%s failed executing, please point LLVM_CONFIG "
  RuntimeError: llvm-config failed executing, please point LLVM_CONFIG to the path for llvm-config
  error: command '/Users/keunwoochoi/miniconda3/envs/omnizart/bin/python' failed with exit code 1
  ERROR: Failed building wheel for llvmlite
  Running setup.py clean for llvmlite
  Building wheel for madmom (setup.py) ... done
  Created wheel for madmom: filename=madmom-0.16.1-cp39-cp39-macosx_10_9_x86_64.whl size=20034572 sha256=9c8033540055c2adbd97b32cc4b9d2162353c87c42c70833e96bf910fc44e7bf
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/36/ef/d2/83e70102b6dd138c8caa880441130bed9618ce53bd8323e339
  Building wheel for pretty-midi (setup.py) ... done
  Created wheel for pretty-midi: filename=pretty_midi-0.2.9-py3-none-any.whl size=5591952 sha256=8c057d76530e28dc5a73d08425224ab7fe4031ec2c1ad6cc6a9f5c30ae9286b7
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/f0/3e/3d/126d7680aae66f3baed077a43a77b91f86e1b4a887fd703cc6
  Building wheel for pyrsistent (setup.py) ... done
  Created wheel for pyrsistent: filename=pyrsistent-0.17.3-cp39-cp39-macosx_10_9_x86_64.whl size=66664 sha256=c29100d99826858bc823131f85e019a5c6059beba6c5977f9d415b7fd47b246f
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/00/cc/12/e712d3979a4aeee423b34a832d97c803e1a67659060315a7e0
  Building wheel for resampy (setup.py) ... done
  Created wheel for resampy: filename=resampy-0.2.2-py3-none-any.whl size=320718 sha256=30f1ac0e52485a15054e353b0e26b6e3f7c7d23d0f9df09db0ee7c8c3ed413f6
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/86/2c/7d/46a32a246b0e5939cea2c5ec1492164073e0c5d16d666ae2cd
  Building wheel for termcolor (setup.py) ... done
  Created wheel for termcolor: filename=termcolor-1.1.0-py3-none-any.whl size=4829 sha256=75f7097a55c2ff3dbd7a667e314e49ead82b30fe639dcba46bbd05261a854d39
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/b6/0d/90/0d1bbd99855f99cb2f6c2e5ff96f8023fad8ec367695f7d72d
  Building wheel for vamp (setup.py) ... done
  Created wheel for vamp: filename=vamp-1.1.0-cp39-cp39-macosx_10_9_x86_64.whl size=121246 sha256=de70095b604a0458d0acee1ac3e47eeeddc509b8d3568e38c13cc4e329e22149
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/b4/05/84/e80fa019e912618654dbfe89cbc2614701dcd7fbbf744aa012
  Building wheel for wrapt (setup.py) ... done
  Created wheel for wrapt: filename=wrapt-1.12.1-cp39-cp39-macosx_10_9_x86_64.whl size=30854 sha256=1580e6613eb95ce043f8580850b2893590216f18ee261fa3c2e9cbbda072ffe2
  Stored in directory: /Users/keunwoochoi/Library/Caches/pip/wheels/98/23/68/efe259aaca055e93b08e74fbe512819c69a2155c11ba3c0f10
Successfully built numba audioread madmom pretty-midi pyrsistent resampy termcolor vamp wrapt
Failed to build llvmlite
Installing collected packages: urllib3, pyasn1, idna, chardet, six, rsa, requests, pyasn1-modules, oauthlib, cachetools, requests-oauthlib, pyparsing, pycparser, numpy, llvmlite, google-auth, werkzeug, threadpoolctl, tensorboard-plugin-wit, tensorboard-data-server, scipy, protobuf, packaging, numba, markdown, joblib, grpcio, google-auth-oauthlib, cffi, appdirs, absl-py, wrapt, typing-extensions, termcolor, tensorflow-estimator, tensorboard, soundfile, scikit-learn, resampy, pyrsistent, pooch, opt-einsum, mido, keras-preprocessing, keras-nightly, h5py, google-pasta, gast, flatbuffers, decorator, audioread, attrs, astunparse, vamp, tqdm, tensorflow, pyyaml, pyfluidsynth, pretty-midi, opencv-python, madmom, librosa, jsonschema, click, omnizart
  Attempting uninstall: numpy
    Found existing installation: numpy 1.21.0
    Uninstalling numpy-1.21.0:
      Successfully uninstalled numpy-1.21.0
    Running setup.py install for llvmlite ... error
     command: /Users/keunwoochoi/miniconda3/envs/omnizart/bin/python -u -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/setup.py'"'"'; __file__='"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-record-xr4wnbyg/install-record.txt --single-version-externally-managed --compile --install-headers /Users/keunwoochoi/miniconda3/envs/omnizart/include/python3.9/llvmlite
         cwd: /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/
    Complete output (29 lines):
    running install
    running build
    got version from file /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-zm0_lgqc/llvmlite_3f00322f7ef34eaea1d20083d31c6732/llvmlite/_version.py {'version': '0.31.0', 'full': 'fe7d985f6421d87f613bd414479d29d912771562'}
    running build_ext
[…]
