### 2021-06-27 Music-and-Culture-Technology-Lab/omnizart#26 @keunwoochoi
On the other hand, `pip install omnizart` on Python 3.8.10 worked like a charm.
