### 2021-06-27 Music-and-Culture-Technology-Lab/omnizart#26 @keunwoochoi
On Python 3.7, it has similar issues mentioned above.
```

(omnizart38) keunwoochoi@KCs-qmul-mbp Downloads $ conda activate omnizart37
(omnizart37) keunwoochoi@KCs-qmul-mbp Downloads $ python --version
Python 3.7.10
(omnizart37) keunwoochoi@KCs-qmul-mbp Downloads $ pip install omnizart
Collecting omnizart
  Using cached omnizart-0.4.1-py3-none-any.whl (3.6 MB)
Collecting pretty_midi<0.3.0,>=0.2.9
  Using cached pretty_midi-0.2.9.tar.gz (5.6 MB)
Collecting pyyaml<6.0.0,>=5.3.1
  Downloading PyYAML-5.4.1-cp37-cp37m-macosx_10_9_x86_64.whl (249 kB)
     |████████████████████████████████| 249 kB 3.6 MB/s
Collecting opencv-python<5.0.0,>=4.4.0
  Downloading opencv_python-4.5.2.54-cp37-cp37m-macosx_10_15_x86_64.whl (43.7 MB)
     |████████████████████████████████| 43.7 MB 2.8 MB/s
Collecting librosa<0.9.0,>=0.8.0
  Using cached librosa-0.8.1-py3-none-any.whl (203 kB)
Collecting numba==0.48
  Downloading numba-0.48.0-cp37-cp37m-macosx_10_9_x86_64.whl (2.0 MB)
     |████████████████████████████████| 2.0 MB 3.5 MB/s
Collecting mir_eval<0.7,>=0.6
  Using cached mir_eval-0.6.tar.gz (87 kB)
Collecting madmom<0.17.0,>=0.16.1
  Using cached madmom-0.16.1.tar.gz (20.0 MB)
     command: /Users/keunwoochoi/miniconda3/envs/omnizart37/bin/python -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-t060fogt/madmom_45991817f0794878809411c1325d7b2f/setup.py'"'"'; __file__='"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-t060fogt/madmom_45991817f0794878809411c1325d7b2f/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-pip-egg-info-0vfz2cm6
         cwd: /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-t060fogt/madmom_45991817f0794878809411c1325d7b2f/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-t060fogt/madmom_45991817f0794878809411c1325d7b2f/setup.py", line 12, in <module>
        from Cython.Build import cythonize, build_ext
    ModuleNotFoundError: No module named 'Cython'
WARNING: Discarding https://files.pythonhosted.org/packages/c7/a3/9f3de3e8068a3606331134d96b84c8db4f7624d6715be8ab3c1f56e6731d/madmom-0.16.1.tar.gz#sha256=64a86a29106e7c4e0c5f4ec96801c2d1a8492db710e4fe27e097da5517d68cb2 (from https://pypi.org/simple/madmom/). Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
Collecting omnizart
  Downloading omnizart-0.4.0-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 244 kB/s
  Downloading omnizart-0.3.4-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 2.7 MB/s
  Downloading omnizart-0.3.3-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 4.0 MB/s
  Downloading omnizart-0.3.2-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 3.6 MB/s
Collecting urllib3<2.0.0,>=1.25.11
  Using cached urllib3-1.26.6-py2.py3-none-any.whl (138 kB)
Collecting omnizart
  Downloading omnizart-0.3.1-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 4.4 MB/s
  Downloading omnizart-0.3.0-py3-none-any.whl (3.6 MB)
     |████████████████████████████████| 3.6 MB 5.6 MB/s
  Downloading omnizart-0.2.0-py3-none-any.whl (3.5 MB)
     |████████████████████████████████| 3.5 MB 2.2 MB/s
  Downloading omnizart-0.1.1-py3-none-any.whl (3.5 MB)
     |████████████████████████████████| 3.5 MB 2.3 MB/s
  Using cached omnizart-0.1.0-py3-none-any.whl (3.5 MB)
ERROR: Cannot install omnizart==0.1.0, omnizart==0.1.1, omnizart==0.2.0, omnizart==0.3.0, omnizart==0.3.1, omnizart==0.3.2, omnizart==0.3.3, omnizart==0.3.4, omnizart==0.4.0 and omnizart==0.4.1 because these package versions have conflicting dependencies.

The conflict is caused by:
    omnizart 0.4.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.4.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.4 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.3 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.2 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.2.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.1.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.1.0 depends on madmom<0.17.0 and >=0.16.1

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
(omnizart37) keunwoochoi@KCs-qmul-mbp Downloads $ pip install Cython
Collecting Cython
  Downloading Cython-0.29.23-cp37-cp37m-macosx_10_9_x86_64.whl (1.9 MB)
     |████████████████████████████████| 1.9 MB 3.0 MB/s
Installing collected packages: Cython
Successfully installed Cython-0.29.23
(omnizart37) keunwoochoi@KCs-qmul-mbp Downloads $ pip install omnizart
Collecting omnizart
  Using cached omnizart-0.4.1-py3-none-any.whl (3.6 MB)
Collecting mir_eval<0.7,>=0.6
  Using cached mir_eval-0.6.tar.gz (87 kB)
Collecting opencv-python<5.0.0,>=4.4.0
  Using cached opencv_python-4.5.2.54-cp37-cp37m-macosx_10_15_x86_64.whl (43.7 MB)
Collecting jsonschema<4.0.0,>=3.2.0
  Using cached jsonschema-3.2.0-py2.py3-none-any.whl (56 kB)
Collecting click<8.0.0,>=7.1.2
  Using cached click-7.1.2-py2.py3-none-any.whl (82 kB)
Collecting pyfluidsynth<2.0.0,>=1.2.5
  Using cached pyFluidSynth-1.3.0-py3-none-any.whl (18 kB)
Collecting tqdm<5.0.0,>=4.49.0
  Using cached tqdm-4.61.1-py2.py3-none-any.whl (75 kB)
Collecting urllib3==1.26.4
  Using cached urllib3-1.26.4-py2.py3-none-any.whl (153 kB)
Collecting madmom<0.17.0,>=0.16.1
  Using cached madmom-0.16.1.tar.gz (20.0 MB)
     command: /Users/keunwoochoi/miniconda3/envs/omnizart37/bin/python -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-p82tfqrm/madmom_0ac816fddfc5450b94bdbed40db56a87/setup.py'"'"'; __file__='"'"'/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-p82tfqrm/madmom_0ac816fddfc5450b94bdbed40db56a87/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-pip-egg-info-cecl_u09
         cwd: /private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-p82tfqrm/madmom_0ac816fddfc5450b94bdbed40db56a87/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/b_/qpqwldc90rd6kd6spwd9f_q80000gn/T/pip-install-p82tfqrm/madmom_0ac816fddfc5450b94bdbed40db56a87/setup.py", line 15, in <module>
        import numpy as np
    ModuleNotFoundError: No module named 'numpy'
WARNING: Discarding https://files.pythonhosted.org/packages/c7/a3/9f3de3e8068a3606331134d96b84c8db4f7624d6715be8ab3c1f56e6731d/madmom-0.16.1.tar.gz#sha256=64a86a29106e7c4e0c5f4ec96801c2d1a8492db710e4fe27e097da5517d68cb2 (from https://pypi.org/simple/madmom/). Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
Collecting omnizart
  Using cached omnizart-0.4.0-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.3.4-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.3.3-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.3.2-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.3.1-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.3.0-py3-none-any.whl (3.6 MB)
  Using cached omnizart-0.2.0-py3-none-any.whl (3.5 MB)
  Using cached omnizart-0.1.1-py3-none-any.whl (3.5 MB)
  Using cached omnizart-0.1.0-py3-none-any.whl (3.5 MB)
ERROR: Cannot install omnizart==0.1.0, omnizart==0.1.1, omnizart==0.2.0, omnizart==0.3.0, omnizart==0.3.1, omnizart==0.3.2, omnizart==0.3.3, omnizart==0.3.4, omnizart==0.4.0 and omnizart==0.4.1 because these package versions have conflicting dependencies.

The conflict is caused by:
    omnizart 0.4.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.4.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.4 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.3 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.2 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.3.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.2.0 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.1.1 depends on madmom<0.17.0 and >=0.16.1
    omnizart 0.1.0 depends on madmom<0.17.0 and >=0.16.1

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
(omnizart37) keunwoochoi@KCs-qmul-mbp Downloads $
```
