### 2021-06-27 Music-and-Culture-Technology-Lab/omnizart#26 @keunwoochoi
I'll continue test its functionalities etc on Python 3.8.
