### 2021-06-29 Music-and-Culture-Technology-Lab/omnizart#25 @BreezeWhite
Hi, sorry for the late update.It's been busy these days.

For issue 2-5: We decide to rewrite the whole "Statement of need" section as can been seen in the updated version.

For L56-58:  We understand your concern. Following the definition in the original paper, the term “instrument agnostic” means that the instrument classes in the input can be “uninformed” and we truly did not mean that the model can predict “unseen” instrument classes. To avoid misleading, we guide the readers to read the [reference paper](https://drive.google.com/file/d/1q9SvH7Ozo3l3cEIFJdR-5kfjY_0j6SbE/view?usp=sharing)

For L69-L70: The system for vocal transcription contains a pitch extraction and a note segmentation module.

Thanks for pointing out the above issues. Please check if there are still needs to modify.
