### 2021-06-29 Music-and-Culture-Technology-Lab/omnizart#26 @BreezeWhite
Hi, thanks for issuing the installation bug. I'll try to reproduce the errors and fix the bugs.
