### 2021-06-30 e0d6f05
Modification according to review feedback #25
