+++ paper.md
@@ -38,7 +38,8 @@ bibliography: paper.bib
 
 We present and release Omnizart, a new Python library that provides a streamlined solution to automatic music transcription (AMT).
 Omnizart encompasses modules that construct the life-cycle of deep learning-based AMT, and is designed for ease of use with a compact command-line interface.
-To the best of our knowledge, Omnizart is the first transcription toolkit which offers models covering a wide class of instruments ranging from solo, instrument ensembles, percussion instruments to vocal, as well as models for chord recognition and beat/downbeat tracking, two music information retrieval (MIR) tasks highly related to AMT. 
+To the best of our knowledge, Omnizart is the first toolkit which offers transcription models for various music content including piano solo, instrument ensembles, percussion and vocal. Omnizart also supports models for chord recognition and beat/downbeat trancking, which are highly related to AMT.
+
 In summary, Omnizart incorporates:
 
 - Pre-trained models for frame-level and note-level transcription of multiple pitched instruments, vocal melody, and drum events;
@@ -51,14 +52,9 @@ The complete code base, command-line interface, documentation, as well as demo e
 
 # Statement of need
 
-AMT has been one of the core challenges in MIR because of the multifaceted nature of musical signals. 
-Typically, streams of musical notes performed with various instruments overlap with each other and then create a hierarchy of abstraction. This complicates the task to identify the melodic, timbral, and rhythmic attributes of the music. 
+AMT of polyphonic music is a complicated MIR task because the note- melody-, timbre-, and rhythm-level attributes of music are overlapped with each other in music signals. A unified solution of AMT is therefore in eager demand. AMT is also strongly related to other MIR tasks such as source separation and music generation with transcribed data needed as supervisory resources. 
+Omnizart considers multi-instrument transcription and collects several state-of-the-art (SoTA) models for transcribing pitched and percussive instruments, as well as singing voice out of polyphonic music signals. Omnizart is an AMT tool which unifies multiple transcription utilities and enables further productivity. Omnizart can save one's time and labor in generating a massive amount of multi-track MIDI files, which could have a great impact on music production, music generation, education, and musicology research.
 
-While the majority of the previous solution focuses on single-instrument transcription, Omnizart collects several state-of-tha-art (SoTA) models for transcribing multiple pitched and percussive instruments, as well as vocal out of the interference with rich music polyphony.
-Omnizart also finds it applicability for chord recognition and beat tracking.
-As such, the proposed library offers a unified solution to music transcription for multi-track and modalities.
-
-In short conclusion, Omnizart represents an AMT tool which unifies multiple transcription utilities and enables further productivity. Omnizart can save one's time and labor in generating massive amount of multi-track MIDI files, which could have a great impact on music production, music generation, education, and musicology research.
 
 # Implementation Details
 
@@ -74,12 +70,12 @@ The model achieves 72.50\% and 79.57\% for frame- and note-level F1-scores, resp
 ## Multi-instrument polyphonic transcription
 
 The multi-instrument transcription model extends the piano solo model to support 11 output classes, namely piano, violin, viola, cello, flute, horn, bassoon, clarinet, harpsichord, contrabass, and oboe, accessed from MusicNet  [@thickstun2017learning].
-Notably, the model allows for \emph{instrument-agnostic transcription} where the instruments to transcribe are unknown during inference  [@wu2020multi].
+Detailed characteristics of the model can be seen in [@wu2020multi].
 The evaluation on the test set from MusicNet  [@thickstun2018invariances] yields 66.59\% for the note streaming task.
 
 ## Drum transcription
 
[…]
