### 2021-06-30 bdbdac0
Add details about enabling auto completion
