### 2021-06-30 cf8dff0
Add details about enabling auto completion
