### 2021-07-10 b0950a1
Add more instructions on installation and fix page arrangement
