+++ docs/source/index.rst
@@ -72,7 +72,7 @@ Note-level vocal transcription
    </audio>
 
 
-Frame-level vocal transcription*
+Frame-level vocal transcription
 
 .. raw:: html
 
@@ -137,11 +137,3 @@ References
 
 .. bibliography::
    refs.bib
-
-[1]_ Tsung-Ping Chen and Lis Su, "Harmony Transformer: Incorporating Chord Segmentation Into Harmony Recognition,"
-International Society of Music Information Retrieval Conference (ISMIR), 2019.
-
-[2]_ 
-
-[3]_ Wei-Tsung Lu and Li Su, "Vocal Melody Extraction With Semantic Segmentation and Audio-Symbolic Domain
-Transfer Learning," International Society of Music Information Retrieval Conference (ISMIR), 2018.
