### 2021-07-24 6944efc
Update details of installation steps
