### 2021-07-30 Music-and-Culture-Technology-Lab/omnizart#26 @BreezeWhite
Oh, sorry for the late update. There is another similar issue that I've replied [here](https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/29#issuecomment-877557801). In short, you need to manually install the missing dependencies before installing Omnizart.
