### 2021-08-10 Music-and-Culture-Technology-Lab/omnizart#37 @hagenw
make install fails
I first started by running (inside my virtual environment that I used before to run `omnizart`):

```bash
$ poetry install
$ python -m pytest
```
which results in errors as the checkpoint files are stored under `~/.local/.../site-packages/` as described in `CONTRIBUTING.md`.

I then switched to start from scratch by deactivating my virtual environment and running:

```bash
$ make install
```

but this fails with:

```
Installing the current project: omnizart (0.4.1)
./scripts/install.sh: line 109: omnizart: command not found
Makefile:57: recipe for target 'install' failed
make: *** [install] Error 127
```
