### 2021-08-11 Music-and-Culture-Technology-Lab/omnizart#38 @keunwoochoi
A similar issue with chord transcription, also on MacOS.
```bash
RuntimeError: module compiled against API version 0xe but this version of numpy is 0xd
Traceback (most recent call last):
  File "/Users/admin/miniconda3/envs/omnizart/bin/omnizart", line 8, in <module>
    sys.exit(entry())
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 829, in __call__
    return self.main(*args, **kwargs)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 782, in main
    rv = self.invoke(ctx)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 1066, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/cli/chord/transcribe.py", line 16, in transcribe
    chord.app.transcribe(input_audio, model_path=model_path, output=output)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/utils.py", line 328, in __getattr__
    module = self._load()
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/utils.py", line 313, in _load
    module = importlib.import_module(self.__name__)
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/importlib/__init__.py", line 127, in import_module
  File "<frozen importlib._bootstrap>", line 1006, in _gcd_import
  File "<frozen importlib._bootstrap>", line 983, in _find_and_load
  File "<frozen importlib._bootstrap>", line 967, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 677, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 728, in exec_module
  File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/chord/__init__.py", line 32, in <module>
    from omnizart.chord.app import ChordTranscription
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/chord/app.py", line 14, in <module>
    from omnizart.feature.chroma import extract_chroma
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/omnizart/feature/chroma.py", line 1, in <module>
    import vamp
  File "/Users/admin/miniconda3/envs/omnizart/lib/python3.7/site-packages/vamp/__init__.py", line 163, in <module>
    import vampyhost
ImportError: numpy.core.multiarray failed to import
```
