### 2021-08-11 Music-and-Culture-Technology-Lab/omnizart#38 @keunwoochoi
This is my pip freeze. Python 3.7.10.
```
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
Cython==0.29.24
decorator==5.0.9
ffmpeg-python==0.2.0
future==0.18.2
gast==0.3.3
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h11==0.12.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpcore==0.12.3
httpx==0.16.1
hyperframe==5.2.0
idna==3.2
importlib-metadata==3.10.1
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
librosa==0.8.0
llvmlite==0.31.0
madmom==0.16.1
Markdown==3.3.4
mido==1.2.10
mir-eval==0.6
norbert==0.2.1
numba==0.48.0
numpy==1.18.5
oauthlib==3.1.1
omnizart==0.4.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.2
pooch==1.4.0
pretty-midi==0.2.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyFluidSynth==1.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
samidata-pt @ git+ssh://git@code.byted.org/lab-audio/samidata_pt.git@0e78dbbc1d0772bea7d97413afc308ee5521b548
scikit-learn==0.24.2
scipy==1.4.1
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
spleeter==2.2.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
vamp==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
```
