### 2021-08-11 Music-and-Culture-Technology-Lab/omnizart#38 @keunwoochoi
I have no idea how easy it'd be for you to fix this issue. But there're a lot of mac users among researchers/developers. At least you'd want to add some warning and clarify which on platform it works.
