### 2021-08-15 Music-and-Culture-Technology-Lab/omnizart#37 @BreezeWhite
I'm not sure whether you use `conda` or `python-venv` as your virtual environment. The command `poetry install` will create its own virtual environment if there is no `.venv` directory (which can be created by running `python -m venv .venv`). And if you are using `conda`, there could be conflict with `poetry`.

Could you also try running `./script/install.sh` directly, instead of `make install` and see what happens?
