### 2021-08-15 Music-and-Culture-Technology-Lab/omnizart#38 @BreezeWhite
That is strange... The second one looks like the issue related to `vamp` package, which is originally written in C and then ported to python. I'm not sure whether it is related to the problem of python version compatibility, and if so, there is little I can do to this issue though.
And for the first situation, unfortunately, I don't have a Mac to test and debug. So the best approach would be just as you mentioned, remind the user there could be such problem on Mac in the documentation.
