### 2021-08-16 Music-and-Culture-Technology-Lab/omnizart#37 @hagenw
I deleted the `.venv` folder and created it with
```bash
$ virtualenv .venv
```
Then I executed
```bash
$ ./scripts/install.sh
...
Installing the current project: omnizart (0.4.1)
./scripts/install.sh: line 109: omnizart: command not found
```
