### 2021-08-16 Music-and-Culture-Technology-Lab/omnizart#37 @hagenw
I continued by running the following commands manually:

```bash
$ source .venv/bin/activate
$ omnizart download-checkpoints
```

But if I then start running the tests, they still fail, because `pytest` is not installed.
