### 2021-08-19 Music-and-Culture-Technology-Lab/omnizart#38 @keunwoochoi
Sounds good to me. Once the documentation is updated, I'll close this issue!
