+++ CONTRIBUTING.md
@@ -31,7 +31,7 @@ Feedback: what a big trap there is...
 
 ## Documentation
 Automatically generate documents from inline docstrings of module, class, and function. 
-[Hosted document page](http://140.109.21.96:8000/build/html/index.html)
+[Hosted document page](https://music-and-culture-technology-lab.github.io/omnizart-doc/index.html)
 
 Documentation style: Follows `numpy` document flavor. Learn more from [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html).
 
