### 2021-08-19 #3391 @faroit
@keunwoochoi thanks for your update.

> i couldn't finish the review for the performance part as i do not have a direct access to any linux machine that i can easily use. the rest of the checklist items seems to be done.

Is this related to https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/38 ?

If yes, I think it would be unfortunate to require a Linux machine to test the performance of the package. As a mac user myself, I will give it a try and try to help fixing the issue.
