### 2021-08-20 #3391 @hagenw
I'm nearly finished with the review.

The only point missing is getting the repository setup for development following the `CONTRIBUTING` guidelines (which has not worked so far for me, see https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/37), and then trying to run the tests locally and building the documentation locally.
