### 2021-08-20 Music-and-Culture-Technology-Lab/omnizart#37 @BreezeWhite
The first problem is caused by the existence of `.venv` directory, which will be automatically used by `poetry` while installing everything. And this behavior contradicts the original intention that by directly executing `./scripts/install.sh` without arguments, it should install dependencies on the system and will not activate the virtualenv during installation. This will broken at the line 109, which the `omnizart` is installed under `.venv`, but the installation is not using that virtualenv. The bug could be fixed by modifying the installation script I think, by adding `poetry config virtualenvs.in-project false` to line 54.

And for the missing package of `pytest`, it is defined as a development dependency, which should be manually installed by running `poetry install`. This is indicated in the last line of section `Download and install` in `CONTRIBUTING.md`.
