### 2021-08-20 Music-and-Culture-Technology-Lab/omnizart#37 @hagenw
So the problem is that I named the virtual environment `.venv`?

I'm still wondering what a proper solution should/could be as the indicated way in `CONTRIBUTING.md` is not working for me.
