### 2021-08-20 Music-and-Culture-Technology-Lab/omnizart#37 @BreezeWhite
According to the document of poetry [here](https://python-poetry.org/docs/configuration/#virtualenvsin-project), yes.

You mean `make install` not working?
