### 2021-08-20 Music-and-Culture-Technology-Lab/omnizart#37 @hagenw
I started now from scratch.

I made sure that no `.venv` folder existed and no virtual environment was activated and then run:
```bash
$ make install
$ poetry install
$ source .venv/bin/activate
$ python -m pytest
```
And it worked.

So, I guess the guide in `CONTRIBUTING.md` is correct.
Will close this issue.
