### 2021-08-20 Music-and-Culture-Technology-Lab/omnizart#40 @hagenw
Test fails with missing ./tests/resource/sample_feature.hdf file
I run the tests locally and got:

```bash
$ python -m pytest
===================================================================================== test session starts =====================================================================================
rootdir: /home/audeering.local/hwierstorf/tmp/omnizart/omnizart
plugins: mock-3.6.1, cov-2.12.0
collected 81 items

tests/test_load_label.py .......                                                                                                                                                        [  8%]
tests/test_utils.py ..................                                                                                                                                                  [ 30%]
tests/beat/test_app.py ..                                                                                                                                                               [ 33%]
tests/chord/test_app.py ...                                                                                                                                                             [ 37%] tests/drum/test_app.py ..                                                                                                                                                               [ 39%]
tests/drum/test_labels.py .                                                                                                                                                             [ 40%]
tests/drum/test_prediction.py ............                                                                                                                                              [ 55%]
tests/feature/test_wrapper_func.py .F                                                                                                                                                   [ 58%]
tests/music/test_app.py ....                                                                                                                                                            [ 62%]
tests/music/test_inference.py .....                                                                                                                                                     [ 69%]
tests/music/test_labels.py ........                                                                                                                                                     [ 79%]
tests/music/test_prediction.py ..........                                                                                                                                               [ 91%]
tests/patch_cnn/test_app.py ..                                                                                                                                                          [ 93%]
tests/vocal/test_app.py ...                                                                                                                                                             [ 97%]
tests/vocal_contour/test_app.py ..                                                                                                                                                      [100%]

========================================================================================== FAILURES ===========================================================================================
___________________________________________________________________________________ test_extract_patch_cqt ____________________________________________________________________________________

mocker = <pytest_mock.plugin.MockerFixture object at 0x7f95824cb0b8>

    def test_extract_patch_cqt(mocker):
>       with h5py.File("./tests/resource/sample_feature.hdf", "r") as fin:

tests/feature/test_wrapper_func.py:14:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
.venv/lib/python3.6/site-packages/h5py/_hl/files.py:408: in __init__
    swmr=swmr)
.venv/lib/python3.6/site-packages/h5py/_hl/files.py:173: in make_fid
    fid = h5f.open(name, flags, fapl=fapl)
h5py/_objects.pyx:54: in h5py._objects.with_phil.wrapper
    ???
h5py/_objects.pyx:55: in h5py._objects.with_phil.wrapper
    ???
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

>   ???
E   OSError: Unable to open file (unable to open file: name = './tests/resource/sample_feature.hdf', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)

h5py/h5f.pyx:88: OSError
====================================================================================== warnings summary =======================================================================================
.venv/lib/python3.6/site-packages/tensorflow/python/autograph/utils/testing.py:21
  /home/audeering.local/hwierstorf/tmp/omnizart/omnizart/.venv/lib/python3.6/site-packages/tensorflow/python/autograph/utils/testing.py:21: DeprecationWarning: the imp module is deprecated in
 favour of importlib; see the module's documentation for alternative uses
    import imp

.venv/lib/python3.6/site-packages/jsonschema/validators.py:928
tests/test_utils.py::test_serializable_with_schema
tests/test_utils.py::test_serializable_with_schema
tests/music/test_app.py::test_load_model[None]
tests/music/test_app.py::test_load_model[Piano]
tests/music/test_app.py::test_load_model[Stream]
tests/music/test_app.py::test_load_model[Pop]
  /home/audeering.local/hwierstorf/tmp/omnizart/omnizart/.venv/lib/python3.6/site-packages/jsonschema/validators.py:928: DeprecationWarning: The metaschema specified by $schema was not found.
 Using the latest draft to validate, but this will raise an error in the future.
    cls = validator_for(schema)

-- Docs: https://docs.pytest.org/en/latest/warnings.html
=================================================================================== short test summary info ===================================================================================
FAILED tests/feature/test_wrapper_func.py::test_extract_patch_cqt - OSError: Unable to open file (unable to open file: name = './tests/resource/sample_feature.hdf', errno = 2, error messag...
===================================================================== 1 failed, 80 passed, 8 warnings in 61.78s (0:01:01) =====================================================================
```
