### 2021-08-22 5bce11d
Add target for handling installation of dev dependencies
