### 2021-08-22 fbe540f
Update checkpoint size
