### 2021-08-22 9baa299
Update explaination about installing dev packages
