### 2021-08-22 Music-and-Culture-Technology-Lab/omnizart#40 @BreezeWhite
I have updated the script. Now you can execute `make install-dev` to install all the necessary packages and resources now.
