### 2021-08-22 Music-and-Culture-Technology-Lab/omnizart#38 @faroit
> It seems like this issue can be fixed simply by reinstalling the `numpy` package according to the answer [here](https://stackoverflow.com/questions/66060487/valueerror-numpy-ndarray-size-changed-may-indicate-binary-incompatibility-exp), which also worked on my machine. Though the requirement by `spleeter 2.2.2` and `tensorflow 2.3.0` is `numpy<1.19.0`, which conflicts with this re-installation.

that doesn't fix my issue. I tried reinstalling several versions of numpy <1.19.0
