### 2021-08-22 Music-and-Culture-Technology-Lab/omnizart#38 @BreezeWhite
It worked for me by upgrading `numpy` to the newest version and ignoring the requirement errors. It could still execute without any further errors.
