### 2021-08-22 Music-and-Culture-Technology-Lab/omnizart#38 @faroit
also no luck using docker

```
2021-08-22 12:57:33.615485: F tensorflow/core/lib/monitoring/sampler.cc:42] Check failed: bucket_limits_[i] > bucket_limits_[i - 1] (0 vs. 10)
qemu: uncaught target signal 6 (Aborted) - core dumped
Aborted
```
