### 2021-08-22 Music-and-Culture-Technology-Lab/omnizart#38 @faroit
> It worked for me by upgrading `numpy` to the newest version and ignoring the requirement errors. It could still execute without any further errors.

tried that as well. same error. As I said, I will try on an Intel Mac later. It looks like tf 2.3 doesn't run on ARM macos at all. Maybe you should note that in the readme
