### 2021-08-23 Music-and-Culture-Technology-Lab/omnizart#40 @hagenw
Thanks.

I deleted `.venv` and started using it, but it fails with:

```bash
$ make install
$ make install-dev
...
Installing the current project: omnizart (0.4.1)
Traceback (most recent call last):
  File "./scripts/download_dev_resource.sh", line 5, in <module>
    from omnizart.remote import download_large_file_from_google_drive
ModuleNotFoundError: No module named 'omnizart'
Makefile:61: recipe for target 'install-dev' failed
make: *** [install-dev] Error 1
```
