### 2021-08-23 Music-and-Culture-Technology-Lab/omnizart#38 @faroit
Okay, I can verify it on an intel mac now. I can also confirm that it runs fine when upgrading to numpy>=1.20.0
