### 2021-08-23 Music-and-Culture-Technology-Lab/omnizart#38 @faroit
@BreezeWhite the issue is related to https://github.com/CPJKU/madmom/pull/474 and is already fixed in https://github.com/CPJKU/madmom/pull/474 I suggest you update to that version of madmom ones its released and close the issue then.
