### 2021-08-23 #3391 @faroit
@keunwoochoi your issue with macos seems to be caused by madmom. While the issues is being addressed, please check https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/38 for a workaround to run the software so that we can proceed with the review quickly.
