### 2021-08-23 #3391 @faroit
@hagenw

> The only point missing is getting the repository setup for development following the `CONTRIBUTING` guidelines (which has not worked so far for me, see [Music-and-Culture-Technology-Lab/omnizart#37](https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/37)), and then trying to run the tests locally and building the documentation locally.

seems like that this is solved. Please close the remaining checkboxes if this is the case. Thanks a lot for your review!
