### 2021-08-23 #3391 @hagenw
I'm still not able to run the tests locally as the updated installation instruction does not work for me, see https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/40#issuecomment-903539742

But as I will be away for two weeks from tomorrow it might be fine if I just tick the boxes or somebody else can try to get the tests running locally.
