### 2021-08-23 ad58650
Use another approach for grabbing dev resource from the cloud
