### 2021-08-23 Music-and-Culture-Technology-Lab/omnizart#38 @BreezeWhite
@faroit thanks for the update. I will keep track on that issue. I will also update the README mentioning that Omnizart is currently not compatible for ARM macos.
