### 2021-08-23 Music-and-Culture-Technology-Lab/omnizart#40 @faroit
@BreezeWhite I have a similar problem with the test vectors.

I was running `omnizart download-checkpoints`and it downloaded the checkpoints to `omnizart/env/lib/python3.8/site-packages/omnizart/checkpoints/...`

however the local tests failed to find them when running `make test`

```
FAILED tests/chord/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/chord/chord_v1/weights. Perhaps not yet downloaded?
FAILED tests/chord/test_app.py::test_load_model[ChordV1] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/chord/chord_v1/weights. Perhaps not yet downl...
FAILED tests/chord/test_app.py::test_load_pb_model[ChordV1] - OSError: SavedModel file does not exist at: /Users/faro/repositories/omnizart/omnizart/checkpoints/chord/chord_v1/{saved_model.pbtxt|saved_mode...
FAILED tests/drum/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/drum/drum_keras/weights.h5. Perhaps not yet downl...
FAILED tests/drum/test_app.py::test_load_model[Keras] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/drum/drum_keras/weights.h5. Perhaps not yet down...
FAILED tests/feature/test_wrapper_func.py::test_extract_patch_cqt - OSError: Unable to open file (unable to open file: name = './tests/resource/sample_feature.hdf', errno = 2, error message = 'No such file...
FAILED tests/music/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/music/music_piano/weights.h5. Perhaps not yet do...
FAILED tests/music/test_app.py::test_load_model[Piano] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/music/music_piano/weights.h5. Perhaps not yet d...
FAILED tests/music/test_app.py::test_load_model[Stream] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/music/music_note_stream/weights.h5. Perhaps no...
FAILED tests/music/test_app.py::test_load_model[Pop] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/music/music_pop/weights.h5. Perhaps not yet downl...
FAILED tests/patch_cnn/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/patch_cnn/patch_cnn_melody/weights.h5. Perha...
FAILED tests/patch_cnn/test_app.py::test_load_model[Melody] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/patch_cnn/patch_cnn_melody/weights.h5. Per...
FAILED tests/vocal/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/vocal/vocal_semi/weights. Perhaps not yet downlo...
FAILED tests/vocal/test_app.py::test_load_model[Semi] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/vocal/vocal_semi/weights. Perhaps not yet downlo...
FAILED tests/vocal/test_app.py::test_load_pb_model[Semi] - OSError: SavedModel file does not exist at: /Users/faro/repositories/omnizart/omnizart/checkpoints/vocal/vocal_semi/{saved_model.pbtxt|saved_model...
FAILED tests/vocal_contour/test_app.py::test_load_model[None] - FileNotFoundError: Checkpoint file not found: /Users/faro/repositories/omnizart/omnizart/checkpoints/vocal/contour/weights.h5. Perhaps not ye...
FAILED tests/vocal_contour/test_app.py::test_load_model[VocalContour] - FileNotFoundError: No checkpoint of vocal-contour found in /Users/faro/repositories/omnizart/omnizart/checkpoints/vocal/contour
===
```
