### 2021-08-24 Music-and-Culture-Technology-Lab/omnizart#38 @keunwoochoi
on my intel mac, it works once i manually upgrade `numpy` to the newest version (1.21.2). @BreezeWhite please update us once you update the madmom version in Omnizart.
