### 2021-08-24 Music-and-Culture-Technology-Lab/omnizart#41 @keunwoochoi
Documentation - specify model for each function exactly
(https://github.com/openjournals/joss-reviews/issues/3391)

For developers who are not MIR researchers, it'd be much clearer if the used model is specified in each API (e.g., https://music-and-culture-technology-lab.github.io/omnizart-doc/music/cli.html) instead of only in the main doc (https://music-and-culture-technology-lab.github.io/omnizart-doc/). And that's because,

> musical notes of instruments [WCS20], chord progression [CS19], drum events [WWS20], frame-level vocal melody [LS18], note-level vocal melody [HS20], and beat [CS20].

this description might not be 100%  clear for non-researchers.
