### 2021-08-24 Music-and-Culture-Technology-Lab/omnizart#41 @keunwoochoi
To clarify; just realized they are specified in the APIs (https://music-and-culture-technology-lab.github.io/omnizart-doc/music/api.html). Maybe specifying which API is being used for each command line interface would do the job.
