### 2021-08-24 Music-and-Culture-Technology-Lab/omnizart#42 @keunwoochoi
Performance issue 
(https://github.com/openjournals/joss-reviews/issues/3391)

In the review checklist, we have this item

>  Functionality: Have the functional claims of the software been confirmed?
>  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

On it, can you improve the doc regarding how models are supposed to perform? I ran a few files with `omnizart music transcribe` and realized for typical multi-instrument pop music, the performance might be under general expectation people would have.

Also, I found "music" to be a little too general. Shall we instead say it like "harmonic part", "tonal instruments", etc, and specify that the output would be a single-track midi file?
