### 2021-09-05 Music-and-Culture-Technology-Lab/omnizart#40 @BreezeWhite
The checkpoints are supposed to be saved to where `omnizart` was cloned, and not being installed via `pip install omnizart`. It seems like `omnizart` was "installed" and not cloned directly from github, and thus the checkpoints were saved to the wrong place.
