### 2021-09-05 Music-and-Culture-Technology-Lab/omnizart#40 @faroit
>  It seems like omnizart was "installed" and not cloned directly from github, and thus the checkpoints were saved to the wrong place.

i installed it via `pip install -e .` I think this would be the preferred way for local development and tests. Can you make sure the paths are adjusted for this installation type?
