### 2021-09-10 Music-and-Culture-Technology-Lab/omnizart#41 @BreezeWhite
Maybe I can add a reference link to the API page if someone wants to know more information about the model.
