### 2021-09-12 Music-and-Culture-Technology-Lab/omnizart#40 @faroit
@BreezeWhite

```sh
git clone ...
cd omnizart
conda create -p .env python=3.8
pip install -U pip setuptools
pip install flake8 pylint pytest pytest-cov pytest-mock
pip install .
pip install numpy==1.21.0
py.test tests
```
