### 2021-09-28 2425b86
Update the version of spleeter.
