+++ pyproject.toml
@@ -37,7 +37,7 @@ tqdm = "^4.49.0"
 pyfluidsynth = "^1.2.5"
 vamp = "^1.1.0"
 urllib3 = "1.26.4"
-spleeter = "^2.0.1"
+spleeter = "^2.2.2"
 mir_eval = "^0.6"
 
 [tool.poetry.dev-dependencies]
