### 2021-09-28 df570a1
Use an alternative source for 'download' badge
