### 2021-09-30 Music-and-Culture-Technology-Lab/omnizart#38 @gergol
I have the same issue on my AMD linux machine (Ubuntu 20.04, AMD Ryzen 3950X processor). With a clean install I get the same error as the OP (with the vocal transcriber) and after manually upgrading numpy to the lates version, I get the same as @keunwoochoi .
