### 2021-09-30 a6689dc
Replace opencv with pillow for feature resizing
