+++ pyproject.toml
@@ -29,7 +29,7 @@ pretty_midi = "^0.2.9"
 librosa = "^0.8.0"
 tensorflow = "^2.3.0"
 madmom = "^0.16.1"
-opencv-python = "^4.4.0"
+pillow = "^8.3.2"
 pyyaml = "^5.3.1"
 jsonschema = "^3.2.0"
 numba = "0.48"
