### 2021-09-30 716da19
Delete outdated file
