### 2021-09-30 6155395
Replace opencv with pillow for feature resizing
