### 2021-10-02 8d4d8f6
Use lazy loader to import extraction functions
