### 2021-10-02 e5c9f40
Convert all checkpoints to TF-model format and simplify code for loading models
