### 2021-10-02 b8b4654
Download checkpoint files from the release
