### 2021-10-02 98bfdf8
Fix unit tests after the change of checkpoints migration
