### 2021-10-13 Music-and-Culture-Technology-Lab/omnizart#40 @BreezeWhite
I think I have figured out the root cause.

If you are using `conda` for generating the virtualenv and install everything inside it, the installed omnizart path will look like `/home/breezewhite/omnizart/.env/lib/python3.8/site-packages`. But when running the `pytest`, since you are testing on the **local** developing project, the module path of omnizart will be resolved as `/home/breezewhite/omnizart/omnizart`, not where you've installed to. There is a variable, `MODULE_PATH`, defined inside `omnizart/__init__.py` that is used for determining where the checkpoint files should be stored, and in this case the `MODULE_PATH` would be `/home/breezewhite/omnizart/.env/lib/python3.8/site-packages`. That is why the `pytest` triggered `FileNotFoundError` since the path being resolved inconsistently.

The way to fix this I can figure out now  is not to use `conda` for virtualenv generation while in developing mode. I would suggest using `venv` instead.
