### 2021-10-13 Music-and-Culture-Technology-Lab/omnizart#40 @faroit
@BreezeWhite that works now, thanks.

Unit tests are still failing for

```
FAILED tests/feature/test_wrapper_func.py::test_extract_patch_cqt - OSError: Unable to open file (unable to open file: name = './tests/resou...
```

and

```
FAILED tests/vocal_contour/test_app.py::test_load_model[VocalContour] - FileNotFoundError: No checkpoint of vocal-contour found in /Users/fa...
```
