### 2021-11-11 #3391 @hagenw
Would be fine with me, my main concern was that the guidelines how to run the tests locally did not work for me.
When this is ok for publication then please go ahead.
