### 2021-11-11 #3391 @keunwoochoi
My issues are small documentation issues that wouldn’t be demanding so much. I’m afraid that could mean they will never be fixed if we decide it’s fine for now.
But, I wouldn’t try to convince others if we decide go ahead. All in all, this is my first time reviewing here and I’m learning, too.
