### 2021-11-14 88ae191
Save the model checkpoint in TF format while training
