+++ omnizart/vocal/app.py
@@ -306,9 +306,7 @@ class VocalTranscription(BaseTranscription):
         logger.info("Constructing callbacks")
         callbacks = [
             tf.keras.callbacks.EarlyStopping(patience=settings.training.early_stop, monitor="val_loss"),
-            tf.keras.callbacks.ModelCheckpoint(
-                jpath(model_save_path, "weights"), save_weights_only=True, monitor="val_loss"
-            )
+            tf.keras.callbacks.ModelCheckpoint(model_save_path, monitor="val_loss")
         ]
         logger.info("Callback list: %s", callbacks)
 
