+++ omnizart/music/app.py
@@ -318,16 +318,12 @@ class MusicTranscription(BaseTranscription):
         model_save_path = jpath(settings.model.save_path, model_name)
         ensure_path_exists(model_save_path)
         write_yaml(settings.to_json(), jpath(model_save_path, "configurations.yaml"))
-        write_yaml(model.to_yaml(), jpath(model_save_path, "arch.yaml"), dump=False)
         logger.info("Model output to: %s", model_save_path)
 
         logger.info("Constructing callbacks")
-        weight_name = "weights.{epoch:02d}-{val_loss:.4f}-{val_accuracy:.4f}.h5"
         callbacks = [
-            tf.keras.callbacks.EarlyStopping(
-                patience=settings.training.early_stop, monitor='val_acc'),
-            tf.keras.callbacks.ModelCheckpoint(
-                jpath(model_save_path, weight_name), save_weights_only=True, monitor='val_accuracy'),
+            tf.keras.callbacks.EarlyStopping(patience=settings.training.early_stop, monitor='val_acc'),
+            tf.keras.callbacks.ModelCheckpoint(model_save_path, monitor='val_accuracy'),
             tf.keras.callbacks.LearningRateScheduler(
                 lambda epoch, lr: lr_scheduler(epoch, lr, update_after=3, dec_every=3, dec_rate=0.25))
         ]
