### 2021-11-14 7c6082a
Update the version of depedencies. Especially upgrade TF to 2.5.0
