+++ setup.py
@@ -26,15 +26,25 @@ packages = \
 package_data = \
 {'': ['*'],
  'omnizart': ['checkpoints/beat/beat_blstm/*',
+              'checkpoints/beat/beat_blstm/variables/variables.index',
               'checkpoints/chord/chord_v1/*',
+              'checkpoints/chord/chord_v1/variables/variables.index',
               'checkpoints/drum/drum_keras/*',
+              'checkpoints/drum/drum_keras/variables/variables.index',
               'checkpoints/music/music_note_stream/*',
+              'checkpoints/music/music_note_stream/variables/variables.index',
               'checkpoints/music/music_piano-v2/*',
+              'checkpoints/music/music_piano-v2/variables/variables.index',
               'checkpoints/music/music_piano/*',
+              'checkpoints/music/music_piano/variables/variables.index',
               'checkpoints/music/music_pop/*',
+              'checkpoints/music/music_pop/variables/variables.index',
               'checkpoints/patch_cnn/patch_cnn_melody/*',
+              'checkpoints/patch_cnn/patch_cnn_melody/variables/variables.index',
               'checkpoints/vocal/contour/*',
+              'checkpoints/vocal/contour/variables/variables.index',
               'checkpoints/vocal/vocal_semi/*',
+              'checkpoints/vocal/vocal_semi/variables/variables.index',
               'defaults/*',
               'resource/vamp/*']}
 
@@ -44,13 +54,12 @@ install_requires = \
  'librosa>=0.8.0,<0.9.0',
  'madmom>=0.16.1,<0.17.0',
  'mir_eval>=0.6,<0.7',
- 'numba==0.48',
  'pillow>=8.3.2,<9.0.0',
  'pretty_midi>=0.2.9,<0.3.0',
  'pyfluidsynth>=1.2.5,<2.0.0',
  'pyyaml>=5.3.1,<6.0.0',
- 'spleeter>=2.2.2,<3.0.0',
- 'tensorflow>=2.3.0,<3.0.0',
+ 'spleeter>=2.3.0,<3.0.0',
+ 'tensorflow>=2.5.0,<3.0.0',
  'tqdm>=4.49.0,<5.0.0',
  'urllib3==1.26.4',
  'vamp>=1.1.0,<2.0.0']
@@ -62,7 +71,7 @@ setup_kwargs = {
     'name': 'omnizart',
     'version': '0.4.1',
     'description': 'Omniscient Mozart, being able to transcribe everything in the music.',
[…]
