### 2021-11-15 dc6e103
Fix checkpoint path bug of vocal contour module
