### 2021-11-15 3c9d615
Fix bugs of unit tests after upgrading TF version
