### 2021-11-15 55ccedd
Modify order of Docker command for better utilizing cache
