### 2021-11-15 d5ea503
Fix chord module bug while loading custom layers
