### 2021-11-19 Music-and-Culture-Technology-Lab/omnizart#40 @hagenw
With version 0779fd5699be6605b9944ab3c5013af3c49f65df of the repo the issue is solved for me.

I did run:

```bash
$ make install
$ make install-dev
$ source .venv/bin/activate
$ python -m pytest
```

as proposed in CONTRIBUTING.md and it worked fine for me.

If you like, you can close this issue.
