### 2021-11-27 dcb97a9
Fix bug of loading custom layers of chord module
