+++ setup.py
@@ -27,7 +27,10 @@ package_data = \
 {'': ['*'],
  'omnizart': ['checkpoints/beat/beat_blstm/*',
               'checkpoints/beat/beat_blstm/variables/variables.index',
-              'checkpoints/chord/chord_v1/*',
+              'checkpoints/chord/chord_v1/configurations.yaml',
+              'checkpoints/chord/chord_v1/configurations.yaml',
+              'checkpoints/chord/chord_v1/saved_model.pb',
+              'checkpoints/chord/chord_v1/saved_model.pb',
               'checkpoints/chord/chord_v1/variables/variables.index',
               'checkpoints/drum/drum_keras/*',
               'checkpoints/drum/drum_keras/variables/variables.index',
@@ -41,8 +44,8 @@ package_data = \
               'checkpoints/music/music_pop/variables/variables.index',
               'checkpoints/patch_cnn/patch_cnn_melody/*',
               'checkpoints/patch_cnn/patch_cnn_melody/variables/variables.index',
-              'checkpoints/vocal/contour/*',
-              'checkpoints/vocal/contour/variables/variables.index',
+              'checkpoints/vocal/vocal_contour/*',
+              'checkpoints/vocal/vocal_contour/variables/variables.index',
               'checkpoints/vocal/vocal_semi/*',
               'checkpoints/vocal/vocal_semi/variables/variables.index',
               'defaults/*',
@@ -59,7 +62,7 @@ install_requires = \
  'pyfluidsynth>=1.2.5,<2.0.0',
  'pyyaml>=5.3.1,<6.0.0',
  'spleeter>=2.3.0,<3.0.0',
- 'tensorflow>=2.5.0,<3.0.0',
+ 'tensorflow==2.5.0',
  'tqdm>=4.49.0,<5.0.0',
  'urllib3==1.26.4',
  'vamp>=1.1.0,<2.0.0']
@@ -71,7 +74,6 @@ setup_kwargs = {
     'name': 'omnizart',
     'version': '0.4.2',
     'description': 'Omniscient Mozart, being able to transcribe everything in the music.',
-    'long_description_content_type': 'text/markdown',
[…]
