+++ requirements.txt
@@ -15,10 +15,10 @@ attrs==21.2.0; python_version >= "2.7" and python_full_version < "3.0.0" or pyth
     --hash=sha256:ef6aaac3ca6cd92904cdd0d83f629a15f18053ec84e6432106f7a4d04ae4f5fb
 audioread==2.1.9; python_full_version >= "3.6.1" and python_version < "3.10" and python_version >= "3.6" \
     --hash=sha256:a3480e42056c8e80a8192a54f6729a280ef66d27782ee11cbd63e9d4d1523089
-cached-property==1.5.2; python_version < "3.8" and python_version >= "3.6" \
+cached-property==1.5.2; python_full_version >= "3.6.1" and python_version < "3.8" and python_version >= "3.6" \
     --hash=sha256:9fa5755838eecbb2d234c3aa390bd80fbd3ac6b6869109bfc1b499f7bd89a130 \
     --hash=sha256:df4f613cf7ad9a588cc381aaf4a512d26265ecebd5eb9e1ba12f1319eb85a6a0
-cachetools==4.2.4; python_version >= "3.5" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") \
+cachetools==4.2.4; python_version >= "3.6" and python_version < "3.10" and python_full_version >= "3.6.1" \
     --hash=sha256:92971d3cb7d2a97efff7c7bb1657f21a8f5fb309a37530537c71b1774189f2d1 \
     --hash=sha256:89ea6f1b638d5a73a4f9226be57ac5e4f399d22770b92355f92dcb0f7f001693
 certifi==2021.10.8; python_full_version >= "3.6.1" and python_version < "3.10" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") \
@@ -126,7 +126,7 @@ cython==0.29.24; python_version >= "2.6" and python_full_version < "3.0.0" or py
     --hash=sha256:5fb977945a2111f6b64501fdf7ed0ec162cc502b84457fd648d6a558ea8de0d6 \
     --hash=sha256:f96411f0120b5cae483923aaacd2872af8709be4b46522daedc32f051d778385 \
     --hash=sha256:cdf04d07c3600860e8c2ebaad4e8f52ac3feb212453c1764a49ac08c827e8443
-dataclasses==0.8; python_version >= "3.6" and python_version < "3.7" \
+dataclasses==0.8; python_version >= "3.6" and python_version < "3.7" and python_full_version >= "3.6.1" \
     --hash=sha256:0201d89fa866f68c8ebd9d08ee6ff50c0b255f8ec63a71c16fda7af82bb887bf \
     --hash=sha256:8479067f342acf957dc82ec415d355ab5edb7e7646b90dc6e2fd1d96ad084c97
 decorator==5.1.0; python_full_version >= "3.6.1" and python_version < "3.10" and python_version >= "3.6" \
@@ -143,10 +143,10 @@ future==0.18.2; python_full_version >= "3.6.1" and python_version < "3.10" \
 gast==0.4.0; python_full_version >= "3.6.1" and python_version < "3.10" \
     --hash=sha256:b7adcdd5adbebf1adf17378da5ba3f543684dbec47b1cda1f3997e573cd542c4 \
     --hash=sha256:40feb7b8b8434785585ab224d1568b857edb18297e5a3047f1ba012bc83b42c1
-google-auth-oauthlib==0.4.6; python_version >= "3.6" \
[…]
