+++ pyproject.toml
@@ -27,7 +27,6 @@ python = ">=3.6.1,<3.9"
 click = "^7.1.2"
 pretty_midi = "^0.2.9"
 librosa = "^0.8.0"
-tensorflow = "^2.5.0"
 madmom = "^0.16.1"
 pillow = "^8.3.2"
 pyyaml = "^5.3.1"
@@ -38,6 +37,7 @@ vamp = "^1.1.0"
 urllib3 = "1.26.4"
 spleeter = "^2.3.0"
 mir_eval = "^0.6"
+tensorflow = "2.5.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
