+++ paper.bib
@@ -47,12 +47,15 @@ pages     = {4052--4061},
 year      = {2018},
 }
 
-@article{Chen2018DeepLabV3,
-title={Encoder-Decoder with Atrous Separable Convolution for Semantic Image Segmentation},
-author={Chen, L.-C. and Zhu, Y. and P. George and S. Florian and A. Hartwig},
-journal={arXiv preprint arXiv:1802.02611},
-year={2018},
-doi={10.1007/978-3-030-01234-2_49}
+@incollection{Chen2018DeepLabV3,
+	doi = {10.1007/978-3-030-01234-2_49},
+	url = {https://doi.org/10.1007%2F978-3-030-01234-2_49},
+	year = 2018,
+	publisher = {Springer International Publishing},
+	pages = {833--851},
+	author = {Liang-Chieh Chen and Yukun Zhu and George Papandreou and Florian Schroff and Hartwig Adam},
+	title = {Encoder-Decoder with Atrous Separable Convolution for Semantic Image Segmentation},
+	booktitle = {Computer Vision {\textendash} {ECCV} 2018}
 }
 
 @article{su2015combining,  
@@ -210,7 +213,8 @@ title={Madmom: A new python audio and music signal processing library},
 author={B{\"o}ck, Sebastian and Korzeniowski, Filip and Schl{\"u}ter, Jan and Krebs, Florian and Widmer, Gerhard},
 booktitle={Proceedings of the 24th ACM international conference on Multimedia},
 pages={1174--1178},
-year={2016}
+year={2016},
+doi={10.1145/2964284.2973795}
 }
 
 @inproceedings{mcfee2015librosa,
