### 2021-12-07 Music-and-Culture-Technology-Lab/omnizart#49 @faroit
JOSS editorial review: bibliography
I found inconsistencies in the bibliography when it comes e.g. Proceedings of ISMIR. I suggest to have all bib entries use:

```
booktitle = {Proceedings of the International Society for Music Information Retrieval Conference ({ISMIR})}
```

https://github.com/openjournals/joss-reviews/issues/3391
