### 2021-12-07 #3391 @faroit
@BreezeWhite The submission looks good now and we can finally move forward 👏 I just did check the paper again for grammar and spellings and created a pull request that you can merge in if you like.

Furthermore I found some inconsistencies in the bibliography that should be fixed as well. (see https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/49)

When these changes are made, the paper is ready, please then please make a new release of the main repo that includes all of the changes that have resulted from the review. Please report the version number here.

Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
