### 2021-12-09 #3391 @BreezeWhite
Hi @faroit, thanks for the update. I've published a new version `v0.5.0` that includes all the reviewed contents. The DOI is `https://zenodo.org/record/5769022`.
I'm not sure what you mean about the *archive title*. Do you mean the release title? And where should I refer author names to the 'archive'?
