### 2021-12-09 892f64a
Add DOI and citation information
