### 2021-12-09 Music-and-Culture-Technology-Lab/omnizart#49 @faroit
@BreezeWhite sorry, but this wasn't related to #48.

- please reopen this issue
- go over the bibliography as suggested above
- close when done
