### 2021-12-09 #3391 @faroit
@BreezeWhite the archive (zenodo) looks good. Please have a look at https://github.com/Music-and-Culture-Technology-Lab/omnizart/issues/49 again before we can proceed with the final accept of this paper.
