### 2021-12-09 3aecfb5
Fix inconsistency in bibliography
