+++ paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{abadi2016tensorflow,
 title={Tensorflow: A system for large-scale machine learning},
 author={Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
-booktitle={12th {USENIX} symposium on operating systems design and implementation ({OSDI} 16)},
+booktitle={{USENIX} symposium on operating systems design and implementation ({OSDI})},
 pages={265--283},
 year={2016}
 }
@@ -29,7 +29,7 @@ author    = {Curtis Hawthorne and
              Douglas Eck},
 title     = {Enabling {F}actorized {P}iano {M}usic {M}odeling and {G}eneration with the {MAESTRO}
              {D}ataset},
-booktitle = {7th International Conference on Learning Representations ({ICLR})},
+booktitle = {International Conference on Learning Representations ({ICLR})},
 year      = {2019},
 }
 
@@ -55,11 +55,11 @@ year      = {2018},
 	pages = {833--851},
 	author = {Liang-Chieh Chen and Yukun Zhu and George Papandreou and Florian Schroff and Hartwig Adam},
 	title = {Encoder-Decoder with Atrous Separable Convolution for Semantic Image Segmentation},
-	booktitle = {Computer Vision {\textendash} {ECCV} 2018}
+	booktitle = {Computer Vision {\textendash} {ECCV}}
 }
 
 @article{su2015combining,  
-author={Li Su and Yi{-}Hsuan Yang},  
+author={Su, Li and Yi{-}Hsuan Yang},  
 journal={IEEE/ACM Transactions on Audio, Speech, and Language Processing}, 
 title={Combining {S}pectral and {T}emporal {R}epresentations for {M}ultipitch {E}stimation of {P}olyphonic {M}usic},
 year={2015},  
@@ -78,7 +78,7 @@ author={Rainer Kelz and
         Sebastian B{\"{o}}ck and
         Andreas Arzt and
         Gerhard Widmer},
-booktitle={Proceedings of the 17th International Society for Music Information Retrieval Conference ({ISMIR})},
+booktitle={Proceedings of the International Society for Music Information Retrieval Conference ({ISMIR})},
 pages={475--481},
 year={2016},
 }
@@ -86,7 +86,7 @@ year={2016},
 @inproceedings{wei2021improving,
 title={Improving automatic drum transcription using large-scale audio-to-midi aligned data},
 author={Wei, I-Chieh and Wu, Chih-Wei and Su, Li},
-booktitle={Proc. ICASSP},
+booktitle={{IEEE} International Conference on Acoustics, Speech and Signal Processing ({ICASSP})},
 year={2021},
 doi={10.1109/icassp39728.2021.9414409}
 }
@@ -101,8 +101,8 @@ doi={10.1109/icassp.2018.8462079}
 }
 
 @INPROCEEDINGS{su2018vocal, 
-author={Su, L.}, 
-booktitle={Proc. ICASSP}, 
+author={Su, Li}, 
+booktitle={{IEEE} International Conference on Acoustics, Speech and Signal Processing ({ICASSP})}, 
 title={Vocal Melody Extraction Using Patch-Based {CNN}}, 
 year={2018}, 
 pages={371-375},
@@ -113,7 +113,7 @@ doi={10.1109/icassp.2018.8462420}
 title={Hierarchical classification networks for singing voice segmentation and transcription},
 author={Fu, Z.-S. and Su, L.},
 year={2019},
-booktitle={Proc. ISMIR}
[…]
