### 2021-12-09 Music-and-Culture-Technology-Lab/omnizart#49 @BreezeWhite
@faroit oops, sorry for that :joy: I've fixed the issue according to your suggestion, and additionally checked other inconsistency in the bibliography. Changes can be viewed from [the commit](https://github.com/Music-and-Culture-Technology-Lab/omnizart/commit/3aecfb5b87c59dc2c6011a35f46834cde98c4127).
