### 2021-12-09 #3391 @danielskatz
@BreezeWhite - I'm the AEiC who will accept and publish this, and while proofreading, I found a number of issues, addressed in https://github.com/Music-and-Culture-Technology-Lab/omnizart/pull/50 - please merge this, or let me know what you disagree with, then we can proceed.
